test: verify-macos-app-launch 회귀 test 를 소스와 같은 seam 으로 분해
12,762줄 단일 verify-macos-app-launch.test.mjs 를 소스 모듈 분해와 짝을
맞춰 scripts/verify-macos-app-launch.<seam>.test.mjs 8개로 나눴다
(webview-env · cli-args · process-lock · relation-marker-validators ·
payload-contract · window-accessibility · visual-evidence ·
evidence-payload). 각 test() 블록은 원본 그대로 옮기고, import 목록만
실제 사용 심볼 기준으로 좁혔다 — eslint no-unused-vars 로 검증.

payload-contract 파일은 validateWebviewVerifyPayload 를 검증하는 test 3개
(그중 하나가 원본에서도 ~9,000줄짜리 단일 test) 를 그대로 담아 여전히
크다 — 소스 쪽 동일 함수가 쪼갤 수 없는 단일 함수인 것과 같은 이유.

분해 전후 `node --test` 통과 수 동일: 61/61 (pass 60/61 → 실패 1건은 소스
분해 중 발견한 별도 버그였고 test 분해와는 무관하게 소스 커밋에서 이미
수정됨).
