+++ scripts/verify-macos-app-launch.payload-contract.test.mjs
@@ -1,55 +1,8 @@
 import assert from "node:assert/strict";
-import fs from "node:fs";
-import os from "node:os";
-import path from "node:path";
 import test from "node:test";
-import * as verifier from "./verify-macos-app-launch.mjs";
 import {
-  buildAccessibilityWindowProbeScript,
-  buildAccessibilityTextProbeSwift,
-  buildForegroundActivationScript,
-  bundlePathConflictWarnings,
-  classifyVisualEvidenceBlocker,
-  collectWindowDiagnostics,
-  createVerifyLock,
-  existingProcessPatterns,
-  expectedRelationLabelAgentGateText,
-  formatWindowDiagnosticsPayload,
-  formatVisualEvidenceHandoffLines,
-  gracefulQuitCommandOptions,
-  gracefulQuitExistingAppCommands,
-  isSelectedRelationAgentGateText,
-  isSelectedRelationPrimaryCopyActionText,
-  normalizeWebviewRoute,
-  parseAccessibilityWindowRows,
-  parseMinWindowSize,
-  parseOnscreenWindows,
-  parseVerifyAppLaunchArgs,
   parseWebviewVerifyPayload,
-  selectedRelationRouteRailTextLeak,
-  validateSelectedRelationCardAttentionLane,
-  validateSelectedRelationCardDensityContract,
-  validateSelectedRelationContextSilhouetteMarkers,
-  validateSelectedRelationIdentityMarkers,
-  validateSelectedRelationEndpointVisibilityMarkers,
-  validateSelectedRelationEndpointRouteMarkers,
-  validateRelationLabelFrameGeometryMarkers,
-  validateTopologyConnectorCacheMarkers,
-  validateSelectedRelationLabelCompactMarkers,
-  topologyDragCompanionVectorTolerance,
-  validateAccessibilityWindowRows,
-  validateAccessibilityText,
-  validateCapturableWindowRows,
-  validateFrontmostAccessibilityRows,
-  validateVisualEvidenceStats,
-  validateWindowRequirements,
   validateWebviewVerifyPayload,
-  webviewVerifyEnvPatch,
-  verifyLockPath,
-  waitForExistingProcessesToExit,
-  waitForWebviewVerifyPayload,
-  visualEvidenceBlockerHint,
-  windowCaptureTargets,
 } from "./verify-macos-app-launch.mjs";
 
 function compactZoomDragRelationLabelMarkers(overrides = {}) {
@@ -69,169 +22,6 @@ function compactZoomDragRelationLabelMarkers(overrides = {}) {
     ...overrides,
   };
 }
-
-test("WebView verification env patch carries route, drag, composer, and requested window size", () => {
-  assert.deepEqual(
-    webviewVerifyEnvPatch({
-      requireWebviewRoute: "/en/topology/",
-      verifyTopologyDrag: true,
-      verifyTopologySelectedRelation: true,
-      verifyTopologyNodePopover: true,
-      verifyTopologyCreateNode: true,
-      verifyTopologyFocusNoop: true,
-      webviewWindowSize: { width: 1100, height: 800 },
-    }),
-    {
-      ONTOLOGY_ATLAS_VERIFY_WEBVIEW: "1",
-      ONTOLOGY_ATLAS_VERIFY_ROUTE: "/en/topology/",
-      ONTOLOGY_ATLAS_VERIFY_TOPOLOGY_DRAG: "1",
-      ONTOLOGY_ATLAS_VERIFY_TOPOLOGY_SELECTED_RELATION: "1",
-      ONTOLOGY_ATLAS_VERIFY_TOPOLOGY_NODE_POPOVER: "1",
-      ONTOLOGY_ATLAS_VERIFY_TOPOLOGY_CREATE_NODE: "1",
-      ONTOLOGY_ATLAS_VERIFY_TOPOLOGY_FOCUS_NOOP: "1",
-      ONTOLOGY_ATLAS_VERIFY_WINDOW_SIZE: "1100x800",
-    },
-  );
-});
-
-test("Tauri node popover verifier captures compact facts even when the inspector starts expanded", () => {
-  const tauriLib = fs.readFileSync("src-tauri/src/lib.rs", "utf8");
-
-  assert.equal(
-    tauriLib.includes('data-node-popover-toggle="collapse"'),
-    true,
-    "expanded node popover verifier needs the existing collapse control",
-  );
-  assert.equal(
-    tauriLib.includes("clicked-collapse-for-compact"),
-    true,
-    "verifier should collapse an already-expanded popover before taking compact facts snapshot",
-  );
-  assert.equal(
-    tauriLib.includes("result.compact?.factsVisible"),
-    true,
-    "verifier should only accept an expanded popover after compact facts were captured",
-  );
-});
-
-test("selected relation label agent gate text exposes MCP and CLI for handoff-ready facts", () => {
-  assert.equal(expectedRelationLabelAgentGateText("handoff-ready"), "MCP/CLI");
-  assert.equal(expectedRelationLabelAgentGateText("preflight-first"), "check");
-  assert.equal(expectedRelationLabelAgentGateText("review-first"), "review");
-});
-
-test("selected relation inspector agent gate text accepts localized proof copy", () => {
-  assert.equal(isSelectedRelationAgentGateText("MCP/CLI ready"), true);
-  assert.equal(isSelectedRelationAgentGateText("Agent gate handoff ready"), true);
-  assert.equal(isSelectedRelationAgentGateText("설명 가능"), true);
-  assert.equal(isSelectedRelationAgentGateText("사전 점검 먼저"), true);
-  assert.equal(isSelectedRelationAgentGateText("검토 먼저"), true);
-  assert.equal(isSelectedRelationAgentGateText("설명"), false);
-  assert.equal(isSelectedRelationAgentGateText(""), false);
-});
-
-test("selected relation primary copy action text accepts compact localized labels", () => {
-  assert.equal(
-    isSelectedRelationPrimaryCopyActionText({
-      text: "설명",
-      action: "explain_relation",
-      locale: "ko",
-    }),
-    true,
-  );
-  assert.equal(
-    isSelectedRelationPrimaryCopyActionText({
-      text: "설명 복사",
-      action: "explain_relation",
-      locale: "ko",
-    }),
-    true,
-  );
-  assert.equal(
-    isSelectedRelationPrimaryCopyActionText({
-      text: "관계 설명 복사",
-      action: "explain_relation",
-      locale: "ko",
-    }),
-    true,
-  );
-  assert.equal(
-    isSelectedRelationPrimaryCopyActionText({
-      text: "점검 복사",
-      action: "relation_check",
-      locale: "ko",
-    }),
-    true,
-  );
-  assert.equal(
-    isSelectedRelationPrimaryCopyActionText({
-      text: "점검",
-      action: "relation_check",
-      locale: "ko",
-    }),
-    true,
-  );
-  assert.equal(
-    isSelectedRelationPrimaryCopyActionText({
-      text: "Copy explain",
-      action: "explain_relation",
-      locale: "en",
-    }),
-    true,
-  );
-  assert.equal(
-    isSelectedRelationPrimaryCopyActionText({
-      text: "설명 가능",
-      action: "explain_relation",
-      locale: "ko",
-    }),
-    false,
-  );
-});
-
-test("selected relation hidden route rail text leak detector catches collapsed chip text", () => {
-  assert.equal(
-    selectedRelationRouteRailTextLeak({
-      bodyText: "CONTAINS ×5 1 STRONGFACTSRCMCP/CLIEXPLAIN",
-    }),
-    true,
-  );
-  assert.equal(
-    selectedRelationRouteRailTextLeak({
-      bodyText: "CONTAINS 6\ncontains 6\nS1MCP/CLI",
-    }),
-    true,
-  );
-  assert.equal(
-    selectedRelationRouteRailTextLeak({
-      bodyText: "CONTAINS ×5 1\nOntology Hub — Mode-Aware\nMCP/CLI",
-    }),
-    false,
-  );
-});
-
-test("topology drag companion vector tolerance scales with wide WebView UI scale", () => {
-  assert.equal(topologyDragCompanionVectorTolerance({}), 8);
-  assert.equal(topologyDragCompanionVectorTolerance({ topologyUiScale: 0.9 }), 8);
-  assert.equal(topologyDragCompanionVectorTolerance({ topologyUiScale: 1.32 }), 10.56);
-  assert.equal(topologyDragCompanionVectorTolerance({ topologyUiScale: 4 }), 14);
-});
-
-test("topology drag verifier captures reactive motion before selecting a relation label", () => {
-  const source = fs.readFileSync(
-    path.resolve("src-tauri/src/lib.rs"),
-    "utf8",
-  );
-  const relationClickIndex = source.indexOf("result.relationLabelClicked = true");
-  const reactiveMotionIndex = source.indexOf("result.dragReactiveMotionLinkedPolicy");
-
-  assert.notEqual(relationClickIndex, -1);
-  assert.notEqual(reactiveMotionIndex, -1);
-  assert.ok(
-    reactiveMotionIndex < relationClickIndex,
-    "drag reactive proof must be captured before relation label click mutates the topology state",
-  );
-});
 
 test("WebView verification requires Add Concept backdrop when the composer is open", () => {
   const payload = {
@@ -910,1138 +700,6 @@ test("WebView verification requires Add Concept backdrop when the composer is op
   );
 });
 
-test("WebView evidence summarizes Add Concept composer blocking proof for agent handoff", () => {
-  assert.equal(typeof verifier.buildWebviewEvidencePayload, "function");
-  const evidence = verifier.buildWebviewEvidencePayload(
-    {
-      href: "tauri://localhost/ko/topology/?p=domain%3Aviews&mode=focus",
-      width: 1512,
-      height: 917,
-      markers: {
-        topologyCreateNodeOpen: true,
-        topologyAttentionWinner: "blocking-composer",
-        topologyCreateNodePanelAttentionRole: "blocking-composer",
-        topologyCreateNodePanelPlacementContract: "centered-blocking-edit",
-        topologyCreateNodeSurfaceRole: "blocking-edit-surface",
-        topologyCreateNodeElevationContract: "solid-panel-over-dimmed-map",
-        topologyCreateNodeSizeContract: "bounded-centered-composer",
-        topologyCreateNodePanelTopToken: "--topology-blocking-composer-top",
-        topologyCreateNodePanelWidthToken: "--topology-blocking-composer-width",
-        topologyCreateNodePanelMaxHeightToken: "--topology-blocking-composer-max-height",
-        topologyCreateNodeFormSurfaceToken: "--topology-blocking-composer-surface",
-        topologyCreateNodeFormBorderToken: "--topology-blocking-composer-border",
-        topologyCreateNodeFormShadowToken: "--topology-blocking-composer-shadow",
-        topologyCreateNodePanelRole: "dialog",
-        topologyCreateNodePanelAriaModal: "true",
-        topologyCreateNodeFocusInside: true,
-        topologyCreateNodeActiveElementTestId: "create-node-title",
-        topologyCreateNodeBackdropVisible: true,
-        topologyCreateNodeBackdropCoversViewport: true,
-        topologyCreateNodeBackdropPointerEvents: "auto",
-        topologyCreateNodeBackdropContract: "blocks-map-and-closes-composer",
-        topologyCreateNodeBackdropSurfaceToken: "--topology-blocking-backdrop-surface",
-        topologyCreateNodeBackdropBackground: "oklab(0 0 0 / 0.68)",
-        topologyCreateNodeBackdropFilter: "none",
-        topologyBlockingComposerOverlayContract: "exclusive-blocking-composer",
-        topologyInteractiveOverlayCount: 1,
-        topologyInteractiveOverlayNames: ["topology-create-node-backdrop"],
-        topologyMapSurfaceBlockingEdit: true,
-        topologyMapSurfaceDemoted: true,
-        topologyMapSurfaceDimOpacity: 0.24,
-        topologyMapSurfaceDimOpacityToken: "--topology-blocking-map-opacity",
-        topologyMapSurfaceFilterToken: "--topology-blocking-map-filter",
-        topologyMapSurfaceInteractionContract: "suppressed-while-blocking-composer",
-        topologyMapSurfacePointerEvents: "none",
-        topologyTransientSurfaceCount: 0,
-        topologyTransientSurfaceNames: [],
-        topologyTransientSurfaceContract: "blocking-surface-wins",
-        topologyCreateNodePanelTop: 142,
-        topologyCreateNodePanelBottom: 584,
-        topologyCreateNodePanelLeft: 476,
-        topologyCreateNodePanelRight: 1036,
-        topologyCreateNodePanelWidth: 560,
-        topologyCreateNodePanelHeight: 442,
-        topologyCreateNodePanelCenterOffset: 0,
-      },
-    },
-    {
-      capturedAt: "2026-06-16T12:00:00.000Z",
-      visualEvidence: {
-        screenshotPath: ".tmp/ontology-atlas-composer-blocking-ko.png",
-        screenshotStatus: "saved",
-        bytes: 733308,
-        method: "window-id",
-      },
-    },
-  );
-
-  assert.equal(evidence.capturedAt, "2026-06-16T12:00:00.000Z");
-  assert.deepEqual(evidence.composerBlockingProof, {
-    proof: "topology-add-concept-composer-blocking",
-    status: "proved",
-    route: "/ko/topology/?p=domain%3Aviews&mode=focus",
-    attention: {
-      winner: "blocking-composer",
-      panelRole: "blocking-composer",
-      placementContract: "centered-blocking-edit",
-      surfaceRole: "blocking-edit-surface",
-      elevationContract: "solid-panel-over-dimmed-map",
-      sizeContract: "bounded-centered-composer",
-      topToken: "--topology-blocking-composer-top",
-      widthToken: "--topology-blocking-composer-width",
-      maxHeightToken: "--topology-blocking-composer-max-height",
-      surfaceToken: "--topology-blocking-composer-surface",
-      borderToken: "--topology-blocking-composer-border",
-      shadowToken: "--topology-blocking-composer-shadow",
-      role: "dialog",
-      ariaModal: "true",
-      focusInside: true,
-      activeElementTestId: "create-node-title",
-    },
-    backdrop: {
-      visible: true,
-      coversViewport: true,
-      pointerEvents: "auto",
-      contract: "blocks-map-and-closes-composer",
-      surfaceToken: "--topology-blocking-backdrop-surface",
-      background: "oklab(0 0 0 / 0.68)",
-      dimAlpha: 0.68,
-      filter: "none",
-    },
-    map: {
-      blockingEdit: true,
-      demoted: true,
-      dimOpacity: 0.24,
-      dimOpacityToken: "--topology-blocking-map-opacity",
-      filterToken: "--topology-blocking-map-filter",
-      interactionContract: "suppressed-while-blocking-composer",
-      pointerEvents: "none",
-    },
-    overlays: {
-      contract: "exclusive-blocking-composer",
-      count: 1,
-      names: ["topology-create-node-backdrop"],
-    },
-    transients: {
-      contract: "blocking-surface-wins",
-      count: 0,
-      names: [],
-      dismissedSurfaceKinds: [
-        "context-menu",
-        "selected-relation",
-        "search-panel",
-        "path-prompt",
-        "node-popover",
-        "support-panel",
-      ],
-      blockingReason: "composer-open",
-    },
-    panel: {
-      top: 142,
-      bottom: 584,
-      left: 476,
-      right: 1036,
-      width: 560,
-      height: 442,
-      centerOffset: 0,
-    },
-    visualSeparation: {
-      status: "proved",
-      attentionLayer: "blocking-composer-over-dimmed-map",
-      scrim: "strong-opaque-scrim",
-      map: "demoted-context-only",
-      panel: "solid-bounded-centered",
-      interaction: "background-blocked",
-      viewport: "14-inch-fullscreen-safe",
-    },
-    agentNextAction: "treat-add-concept-composer-as-current-work-surface",
-    agentHandoff: {
-      currentSurface: "topology-add-concept-composer",
-      mapState: "dimmed-and-interaction-blocked",
-      blockedUntil: "create-or-cancel",
-      visualEvidence: {
-        screenshotPath: path.resolve(".tmp/ontology-atlas-composer-blocking-ko.png"),
-        screenshotStatus: "saved",
-        bytes: 733308,
-        method: "window-id",
-      },
-      nextActions: ["complete-create-node-form", "cancel-composer"],
-    },
-  });
-});
-
-test("WebView evidence summarizes selected relation label handoff proof for agent handoff", () => {
-  const evidence = verifier.buildWebviewEvidencePayload(
-    {
-      href: "tauri://localhost/ko/topology/?p=domain%3Aai-agent-partner&mode=focus",
-      markers: {
-        topologyCardsReady: true,
-        topologyRelationLabelHandoffContract: "label-level-mcp-cli-fallback",
-        topologyRootAttentionWinner: "active-relation-inspector",
-        topologyAgentCurrentSurface: "selected-relation",
-        topologyAgentCurrentSurfaceRole: "active-relation-inspector",
-        topologyAgentCurrentSurfaceRoute:
-          "domain:ai-agent-partner>capability:agent-config-onboarding",
-        topologySelectedNodeId: "",
-        topologySelectedRelationLabelHandoffState: "ready",
-        topologySelectedRelationLabelHandoffGate: "handoff-ready",
-        topologySelectedRelationLabelHandoffPrimaryAction: "explain_relation",
-        topologySelectedRelationLabelHandoffCliFallbackCommand:
-          "ontology-atlas explain 'domain:ai-agent-partner' 'capability:agent-config-onboarding' [vault] --type 'contains'",
-        topologySelectedRelationLabelHandoffFactRoute: "fact>evidence>gate>action",
-        topologySelectedRelationLabelHandoffQuality: "strong",
-        topologySelectedRelationLabelHandoffEvidence: "source-backed",
-        topologySelectedRelationLabelAgentGateKind: "handoff-ready",
-        topologySelectedRelationLabelPrimaryCopyAction: "explain_relation",
-        topologySelectedRelationLabelCliFallbackCommand:
-          "ontology-atlas explain 'domain:ai-agent-partner' 'capability:agent-config-onboarding' [vault] --type 'contains'",
-        topologySelectedRelationLabelFactRoute: "fact>evidence>gate>action",
-        topologySelectedRelationLabelQuality: "strong",
-        topologySelectedRelationLabelEvidenceState: "source-backed",
-        topologySelectedRelationLabelSource: "domain:ai-agent-partner",
-        topologySelectedRelationLabelTarget: "capability:agent-config-onboarding",
-        topologySelectedRelationLabelType: "contains",
-        topologySelectedRelationLabelCount: 4,
-        topologyFocusRelationLabelVisibleText: "포함 ×4 · S1",
-        topologySelectedRelationLabelRoute:
-          "domain:ai-agent-partner>capability:agent-config-onboarding",
-        topologySelectedRelationLabelTypeLabel: "포함 ×4",
-        topologySelectedRelationCardHandoffContract:
-          "selected-relation-card-carries-mcp-cli-fallback",
-        topologySelectedRelationCardHandoffAliasContract:
-          "selected-relation-card-carries-mcp-cli-fallback",
-        topologySelectedRelationCardRoute: "source>target>type>action",
-        topologySelectedRelationCardEndpointRoute:
-          "domain:ai-agent-partner>capability:agent-config-onboarding",
-        topologySelectedRelationCardPrimaryAction: "explain_relation",
-        topologySelectedRelationCardCliFallback:
-          "ontology-atlas explain 'domain:ai-agent-partner' 'capability:agent-config-onboarding' [vault] --type 'contains'",
-        topologySelectedRelationCardSource: "domain:ai-agent-partner",
-        topologySelectedRelationCardTarget: "capability:agent-config-onboarding",
-        topologySelectedRelationCardType: "contains",
-        topologySelectedRelationCardLabelContextContract:
-          "selected-card-preserves-aggregate-label-context",
-        topologySelectedRelationCardLabelCount: 4,
-        topologySelectedRelationCardLabelVisibleText: "포함 ×4 · S1",
-        topologySelectedRelationCardLabelReadableText: "포함 ×4 · S1",
-        topologySelectedRelationCopyPayloadFrom: "domain:ai-agent-partner",
-        topologySelectedRelationCopyPayloadTo: "capability:agent-config-onboarding",
-        topologySelectedRelationHandleStripSource: "domain:ai-agent-partner",
-        topologySelectedRelationHandleStripTarget: "capability:agent-config-onboarding",
-        topologySelectedRelationEndpointRouteContract: "visible-source-target-names-wrap",
-        topologySelectedRelationEndpointRouteWrapPolicy:
-          "wrap-allowed-no-horizontal-overflow",
-        topologySelectedRelationEndpointRouteLineBudget: "2",
-        topologySelectedRelationEndpointRouteSourceName: "AI Agent Partner",
-        topologySelectedRelationEndpointRouteTargetName: "Agent Config Onboarding",
-        topologySelectedRelationEndpointRouteSourceHandle: "domain:ai-agent-partner",
-        topologySelectedRelationEndpointRouteTargetHandle:
-          "capability:agent-config-onboarding",
-        topologySelectedRelationEndpointRouteHandleSummary:
-          "domain:ai-agent-partner → capability:agent-config-onboarding",
-        topologySelectedRelationEndpointRouteText:
-          "AI Agent Partner→Agent Config Onboarding",
-        topologySelectedRelationEndpointRouteReadableText:
-          "AI Agent Partner → Agent Config Onboarding",
-        topologySelectedRelationEndpointReadableRoute:
-          "AI Agent Partner → Agent Config Onboarding",
-        topologySelectedRelationEndpointRouteWidth: 213,
-        topologySelectedRelationEndpointRouteHeight: 30,
-        topologySelectedRelationEndpointRouteClientWidth: 213,
-        topologySelectedRelationEndpointRouteScrollWidth: 213,
-        topologySelectedRelationEndpointVisibilityContract:
-          "selected-relation-keeps-source-target-readable",
-        topologySelectedRelationEndpointExpectedCount: 2,
-        topologySelectedRelationEndpointVisibleCount: 2,
-        topologySelectedRelationEndpointHiddenCount: 0,
-        topologySelectedRelationEndpointCards: [
-          {
-            slug: "domain:ai-agent-partner",
-            roleBadgeText: "FROM",
-            roleBadgeContract: "visible-source-target-role-badge",
-            roleBadgeVisible: true,
-            visible: true,
-            surfaceHidden: "",
-            shift: "safe-shift",
-          },
-          {
-            slug: "capability:agent-config-onboarding",
-            roleBadgeText: "TO",
-            roleBadgeContract: "visible-source-target-role-badge",
-            roleBadgeVisible: true,
-            visible: true,
-            surfaceHidden: "",
-            shift: "safe-shift",
-          },
-        ],
-        topologySelectedRelationContextSilhouettePolicy:
-          "selected-relation-keeps-endpoints-and-orientation-anchors-only",
-        topologySelectedRelationContextSilhouetteActive: true,
-        topologySelectedRelationContextSilhouetteHiddenCount: 8,
-        topologySelectedRelationLowerPriorityVisibleDimmedCount: 0,
-        topologySelectedRelationVisibleOrientationAnchorCount: 3,
-        topologySelectedRelationHiddenContextInteractionContract:
-          "hidden-context-is-not-pointer-focus-or-a11y-target",
-        topologySelectedRelationHiddenContextInteractiveCount: 0,
-        topologyRelationLabelGeometryContract: "frame-positioned-hit-targets",
-        topologyRelationLabelGeometrySource: "after-render-layout-pass",
-        topologyRelationLabelGeometryExpectedCount: 1,
-        topologyRelationLabelGeometryReadyCount: 1,
-        topologyRelationLabelGeometryPendingCount: 0,
-        topologySelectedBlockingSurfaceOverlapActive: true,
-        topologyVisibleCardRectReadPolicy: "frame-state-no-computed-style",
-        topologyVisibleCardSelectedSurfaceRectPolicy:
-          "live-rects-for-postprocess-overlap-safety",
-        topologyVisibleCardRectReadCount: 10,
-        topologyConnectorDomIndexContract: "reuse-card-index",
-        topologyConnectorRectCacheContract: "frame-local-card-rect-cache",
-        topologyConnectorRectCacheFrameFallbackContract:
-          "reuse-card-placement-frame-rects-before-dom-read",
-        topologyConnectorRectCacheAccounting: "reads-plus-hits",
-        topologyConnectorRectCacheSize: 7,
-        topologyConnectorRectCacheSeedCount: 6,
-        topologyConnectorRectCacheReadCount: 0,
-        topologyConnectorRectCacheHitCount: 26,
-        topologyRepositionPassConnectorLabelMs: 0.18,
-        topologyRepositionMaxPassConnectorLabelMs: 0.42,
-        topologyRepositionPassSlowest: "card-placement",
-        topologyRelief: true,
-        topologyCardOverlapCount: 0,
-        topologyCardClippedCount: 0,
-        topologyFixedSurfaceOverlapCount: 0,
-        topologyCardFixedSurfaceOverlapCount: 0,
-        topologyResidualOverlapClearContract:
-          "visibility-cache-proves-selected-surfaces-clear",
-        topologyResidualOverlapReadPolicy: "reuse-visible-card-rect-cache",
-        topologyResidualOverlapClear: true,
-        topologyResidualVisibleCardOverlapCount: 0,
-        topologyResidualFixedSurfaceOverlapCount: 0,
-        topologyResidualCardFixedSurfaceOverlapCount: 0,
-        topologyZoomLensContract: "zoom-in-uses-kind-pins-for-noncritical-context-cards",
-        topologyZoomLensPresentationContract:
-          "camera-or-focus-lens-uses-kind-pins-for-noncritical-context",
-        topologyZoomLensPresentationActive: true,
-        topologyZoomLensPresentationSource: "camera-zoom-in",
-        topologyZoomLensThresholdRatio: 0.98,
-        topologyZoomLensCardCompactionActive: true,
-        topologyZoomLensCameraRatio: 0.74,
-        topologyZoomLensActive: true,
-        topologyZoomLensEligibleCount: 5,
-        topologyZoomLensActiveCardCount: 3,
-          topologyZoomLensPinMinOpacity: 0.42,
-        topologyVisibleCardCount: 2,
-        topologyZoomLensVisibleActiveCardCount: 2,
-        topologyZoomLensFocusEgoReadableContract:
-          "selected-focus-ego-neighbors-stay-readable-in-lens",
-        topologyZoomLensFocusEgoReadableCount: 6,
-        topologyZoomLensPinProximityContract:
-          "zoomed-context-pins-keep-critical-relation-proximity",
-        topologyZoomLensPinProximityActive: true,
-        topologyZoomLensProximityPinCount: 2,
-        topologyZoomLensPinProximityRingToken: "--topology-zoom-lens-pin-proximity-ring",
-          topologyZoomLensPinGlyphContract: "compact-kind-pin-keeps-type-glyph-without-title-card",
-          topologyZoomLensPinGlyphVisibleCount: 2,
-        topologyZoomLensViewportVisibleContract:
-          "visible-lens-pins-match-frame-state",
-        topologyZoomLensPinCanvasContract:
-          "zoom-lens-pins-stay-inside-readable-canvas-safe-band",
-        topologyZoomLensPinCanvasMarginPx: 32,
-        topologyZoomLensPinCanvasClampCount: 4,
-        topologyZoomLensEmptyViewportFallbackContract:
-          "camera-zoom-in-keeps-at-least-one-ontology-mark-visible",
-        topologyZoomLensEmptyViewportFallbackActive: true,
-        topologyZoomLensRelationChromeContract:
-          "camera-zoom-in-demotes-nonselected-relation-chrome",
-        topologyZoomLensRelationChromeActive: true,
-        topologyZoomLensRelationThreadCount: 4,
-        topologyZoomLensRelationLabelSuppressedCount: 2,
-        topologyFocusDetailLensContract:
-          "selected-focus-uses-kind-pins-for-noncritical-ego-context",
-        topologyFocusDetailLensActive: true,
-        topologyFocusDetailConnectorExpressionContract:
-          "focus-detail-lens-demotes-noncritical-connectors",
-        topologyFocusDetailConnectorExpressionActive: true,
-        topologyFocusDetailConnectorExpressionCount: 3,
-        topologySelectedFocusContextRailVisibleContract:
-          "focus-domain-context-rail-reports-visible-and-hidden-cards",
-        topologySelectedFocusContextRailCount: 5,
-        topologySelectedFocusContextRailVisibleCount: 1,
-        topologySelectedFocusContextRailHiddenCount: 4,
-        topologySelectedFocusContextRailHiddenReason: "layout-surface-collision",
-        topologyOverviewDensityLensContract:
-          "zoom-out-overview-uses-kind-pins-for-noncritical-context-cards",
-        topologyOverviewDensityLensThresholdRatio: 1.2,
-        topologyOverviewDensityLensMinWidth: 1800,
-        topologyOverviewDensityLensActive: false,
-        topologyOverviewDensityLensActiveCardCount: 0,
-        topologyOverviewDensityFixedGeographyContract:
-          "overview-density-uses-deterministic-canvas-geography",
-        topologyOverviewDensityFixedGeographyActive: true,
-        topologyOverviewDensityFixedGeographyDragContract:
-          "fixed-overview-geography-disables-card-drag",
-        topologyOverviewDensityFixedGeographyDragLocked: true,
-        topologyOverviewDensityFixedGeographyDragAttempt: "ignored",
-        topologyOverviewDensityFixedGeographySlotCount: 22,
-        topologyOverviewDensityFixedGeographyDomainCount: 6,
-        topologyOverviewDensityFixedGeographyPinCount: 15,
-        topologySupportRailOverlapReadPolicy: "reuse-visible-card-rect-cache",
-        topologyDragActiveOverlapPolicy:
-          "active-cluster-hides-lower-priority-overlaps",
-        topologyDragActiveOverlapReadPolicy: "reuse-visible-card-rect-cache",
-        topologyDragActiveOverlapHiddenCount: 0,
-        topologyFixedSurfaceLiveSuppressionReadPolicy:
-          "reuse-card-placement-frame-rects-before-dom-read",
-        topologyFixedSurfaceLiveSuppressionReadCount: 0,
-        topologyFixedSurfaceLiveSuppressedCount: 0,
-        topologyDragSettleOverlapReadPolicy: "reuse-visible-card-rect-cache",
-      },
-    },
-    { capturedAt: "2026-06-17T12:00:00.000Z" },
-  );
-
-  assert.deepEqual(evidence.agentCurrentSurfaceProof, {
-    proof: "topology-agent-current-surface",
-    status: "proved",
-    route: "/ko/topology/?p=domain%3Aai-agent-partner&mode=focus",
-    attentionWinner: "active-relation-inspector",
-    currentSurface: "selected-relation",
-    currentSurfaceRole: "active-relation-inspector",
-    currentSurfaceRoute: "domain:ai-agent-partner>capability:agent-config-onboarding",
-    selectedNodeId: null,
-    rootSelectedNodeId: null,
-    agentNextAction: "read-selected-relation-surface-before-map-context",
-  });
-  assert.deepEqual(evidence.relationLabelHandoffProof, {
-    proof: "topology-relation-label-handoff",
-    status: "proved",
-    route: "/ko/topology/?p=domain%3Aai-agent-partner&mode=focus",
-    contract: "label-level-mcp-cli-fallback",
-    label: {
-      gate: "handoff-ready",
-      primaryAction: "explain_relation",
-      cliFallback:
-        "ontology-atlas explain 'domain:ai-agent-partner' 'capability:agent-config-onboarding' [vault] --type 'contains'",
-      factRoute: "fact>evidence>gate>action",
-      quality: "strong",
-      evidence: "source-backed",
-      source: "domain:ai-agent-partner",
-      target: "capability:agent-config-onboarding",
-      type: "contains",
-      count: 4,
-      route: "domain:ai-agent-partner>capability:agent-config-onboarding",
-      typeLabel: "포함 ×4",
-    },
-    card: {
-      contract: "selected-relation-card-carries-mcp-cli-fallback",
-      handoffAliasContract: "selected-relation-card-carries-mcp-cli-fallback",
-      route: "source>target>type>action",
-      endpointRoute: "domain:ai-agent-partner>capability:agent-config-onboarding",
-      primaryAction: "explain_relation",
-      cliFallback:
-        "ontology-atlas explain 'domain:ai-agent-partner' 'capability:agent-config-onboarding' [vault] --type 'contains'",
-      source: "domain:ai-agent-partner",
-      target: "capability:agent-config-onboarding",
-      type: "contains",
-      labelContextContract: "selected-card-preserves-aggregate-label-context",
-      labelCount: 4,
-      labelVisibleText: "포함 ×4 · S1",
-      labelReadableText: "포함 ×4 · S1",
-    },
-    root: {
-      attentionWinner: "active-relation-inspector",
-      currentSurface: "selected-relation",
-      currentSurfaceRole: "active-relation-inspector",
-      currentSurfaceRoute: "domain:ai-agent-partner>capability:agent-config-onboarding",
-    },
-    aggregate: {
-      gate: "handoff-ready",
-      primaryAction: "explain_relation",
-      cliFallback:
-        "ontology-atlas explain 'domain:ai-agent-partner' 'capability:agent-config-onboarding' [vault] --type 'contains'",
-      factRoute: "fact>evidence>gate>action",
-      quality: "strong",
-      evidence: "source-backed",
-    },
-    agentNextAction: "run-explain-relation-for-handoff",
-  });
-  assert.deepEqual(evidence.relationEndpointVisibilityProof, {
-    proof: "topology-selected-relation-endpoint-visibility",
-    status: "proved",
-    route: "/ko/topology/?p=domain%3Aai-agent-partner&mode=focus",
-    contract: "selected-relation-keeps-source-target-readable",
-    expectedCount: 2,
-    visibleCount: 2,
-    hiddenCount: 0,
-    source: "domain:ai-agent-partner",
-    target: "capability:agent-config-onboarding",
-    readableRoute: "AI Agent Partner → Agent Config Onboarding",
-    layerReadableRoute: "AI Agent Partner → Agent Config Onboarding",
-    routeProof: {
-      contract: "visible-source-target-names-wrap",
-      wrapPolicy: "wrap-allowed-no-horizontal-overflow",
-      lineBudget: 2,
-      clientWidth: 213,
-      scrollWidth: 213,
-      horizontalOverflow: 0,
-    },
-    cards: [
-      {
-        slug: "domain:ai-agent-partner",
-        visible: true,
-        surfaceHidden: "",
-        shift: "safe-shift",
-      },
-      {
-        slug: "capability:agent-config-onboarding",
-        visible: true,
-        surfaceHidden: "",
-        shift: "safe-shift",
-      },
-    ],
-    agentNextAction: "read-selected-relation-with-source-and-target-cards",
-  });
-  assert.deepEqual(evidence.relationContextSilhouetteProof, {
-    proof: "topology-selected-relation-context-silhouette",
-    status: "proved",
-    route: "/ko/topology/?p=domain%3Aai-agent-partner&mode=focus",
-    policy: "selected-relation-keeps-endpoints-and-orientation-anchors-only",
-    active: true,
-    hiddenCount: 8,
-    lowerPriorityVisibleDimmedCount: 0,
-    visibleOrientationAnchorCount: 3,
-    agentNextAction: "read-selected-relation-before-background-context",
-  });
-  assert.deepEqual(evidence.relationLabelFrameGeometryProof, {
-    proof: "topology-relation-label-frame-geometry",
-    status: "proved",
-    route: "/ko/topology/?p=domain%3Aai-agent-partner&mode=focus",
-    contract: "frame-positioned-hit-targets",
-    source: "after-render-layout-pass",
-    expected: 1,
-    ready: 1,
-    pending: 0,
-    agentNextAction: "trust-frame-positioned-relation-label-hit-targets",
-  });
-  assert.deepEqual(evidence.connectorCacheProof, {
-    proof: "topology-connector-cache-frame-fallback",
-    status: "proved",
-    route: "/ko/topology/?p=domain%3Aai-agent-partner&mode=focus",
-    domIndexContract: "reuse-card-index",
-    cacheContract: "frame-local-card-rect-cache",
-    frameFallbackContract: "reuse-card-placement-frame-rects-before-dom-read",
-    accounting: "reads-plus-hits",
-    size: 7,
-    seedCount: 6,
-    readCount: 0,
-    hitCount: 26,
-    visibleCardClippedCount: 0,
-    agentNextAction: "trust-frame-local-connector-rect-cache-before-reading-labels",
-  });
-  assert.deepEqual(evidence.connectorLabelPassProof, {
-    proof: "topology-connector-label-pass-budget",
-    status: "proved",
-    route: "/ko/topology/?p=domain%3Aai-agent-partner&mode=focus",
-    passMs: 0.18,
-    budgetMs: 3,
-    maxPassMs: 0.42,
-    slowestPass: "card-placement",
-    agentNextAction: "read-relation-labels-after-connector-label-pass-budget",
-  });
-  assert.deepEqual(evidence.visibleCardSelectedSurfaceRectProof, {
-    proof: "topology-visible-card-selected-surface-rect-policy",
-    status: "proved",
-    route: "/ko/topology/?p=domain%3Aai-agent-partner&mode=focus",
-    selectedBlockingSurfaceOverlapActive: true,
-    readPolicy: "frame-state-no-computed-style",
-    selectedSurfaceRectPolicy: "live-rects-for-postprocess-overlap-safety",
-    readCount: 10,
-    agentNextAction: "trust-selected-surface-rect-policy-before-reading-relation",
-  });
-  assert.deepEqual(evidence.residualOverlapProof, {
-    proof: "topology-residual-overlap-clear",
-    status: "proved",
-    route: "/ko/topology/?p=domain%3Aai-agent-partner&mode=focus",
-    visibleCardOverlapCount: 0,
-    visibleCardClippedCount: 0,
-    fixedSurfaceOverlapCount: 0,
-    cardFixedSurfaceOverlapCount: 0,
-    supportRailOverlapReadPolicy: "reuse-visible-card-rect-cache",
-    dragActiveOverlapPolicy: "active-cluster-hides-lower-priority-overlaps",
-    dragActiveOverlapReadPolicy: "reuse-visible-card-rect-cache",
-    dragActiveOverlapHiddenCount: 0,
-    fixedSurfaceLiveSuppressionReadPolicy:
-      "reuse-card-placement-frame-rects-before-dom-read",
-    fixedSurfaceLiveSuppressionReadCount: 0,
-    fixedSurfaceLiveSuppressedCount: 0,
-    dragSettleOverlapReadPolicy: "reuse-visible-card-rect-cache",
-    domMarker: {
-      clearContract: "visibility-cache-proves-selected-surfaces-clear",
-      readPolicy: "reuse-visible-card-rect-cache",
-      clear: true,
-      visibleCardOverlapCount: 0,
-      fixedSurfaceOverlapCount: 0,
-      cardFixedSurfaceOverlapCount: 0,
-    },
-    agentNextAction: "read-relation-surfaces-after-residual-overlap-clear",
-  });
-  assert.deepEqual(evidence.zoomLensProof, {
-    proof: "topology-zoom-lens-kind-pins",
-    status: "proved",
-    route: "/ko/topology/?p=domain%3Aai-agent-partner&mode=focus",
-    contract: "zoom-in-uses-kind-pins-for-noncritical-context-cards",
-    presentationContract: "camera-or-focus-lens-uses-kind-pins-for-noncritical-context",
-    thresholdRatio: 0.98,
-    cardCompactionActive: true,
-    presentationActive: true,
-    presentationSource: "camera-zoom-in",
-    cameraRatio: 0.74,
-    active: true,
-    eligibleCount: 5,
-    activeCardCount: 3,
-    visibleCardCount: 2,
-    visibleActiveCardCount: 2,
-    pinMinOpacity: 0.42,
-    focusEgoReadable: {
-      contract: "selected-focus-ego-neighbors-stay-readable-in-lens",
-      count: 6,
-    },
-    pinGlyph: {
-      contract: "compact-kind-pin-keeps-type-glyph-without-title-card",
-      visibleCount: 2,
-    },
-    proximityPins: {
-      contract: "zoomed-context-pins-keep-critical-relation-proximity",
-      active: true,
-      count: 2,
-      ringToken: "--topology-zoom-lens-pin-proximity-ring",
-    },
-    viewportVisibleContract: "visible-lens-pins-match-frame-state",
-    pinCanvas: {
-      contract: "zoom-lens-pins-stay-inside-readable-canvas-safe-band",
-      marginPx: 32,
-      clampCount: 4,
-    },
-    emptyViewportFallback: {
-      contract: "camera-zoom-in-keeps-at-least-one-ontology-mark-visible",
-      active: true,
-    },
-    relationChrome: {
-      contract: "camera-zoom-in-demotes-nonselected-relation-chrome",
-      active: true,
-      threadCount: 4,
-      labelSuppressedCount: 2,
-    },
-    focusDetail: {
-      contract: "selected-focus-uses-kind-pins-for-noncritical-ego-context",
-      active: true,
-      connectorExpression: {
-        contract: "focus-detail-lens-demotes-noncritical-connectors",
-        active: true,
-        count: 3,
-      },
-      contextRail: {
-        contract: "focus-domain-context-rail-reports-visible-and-hidden-cards",
-        totalCount: 5,
-        visibleCount: 1,
-        hiddenCount: 4,
-        hiddenReason: "layout-surface-collision",
-      },
-    },
-    overviewDensity: {
-      contract: "zoom-out-overview-uses-kind-pins-for-noncritical-context-cards",
-      thresholdRatio: 1.2,
-      minWidth: 1800,
-      active: false,
-      activeCardCount: 0,
-      fixedGeography: {
-        contract: "overview-density-uses-deterministic-canvas-geography",
-        active: true,
-        dragContract: "fixed-overview-geography-disables-card-drag",
-        dragLocked: true,
-        dragAttempt: "ignored",
-        slotCount: 22,
-        domainCount: 6,
-        pinCount: 15,
-      },
-    },
-    agentNextAction: "trust-kind-pin-lens-before-reading-dense-map-cards",
-  });
-});
-
-test("WebView evidence summarizes selected relation visible fact route proof for agent handoff", () => {
-  const evidence = verifier.buildWebviewEvidencePayload(
-    {
-      href: "tauri://localhost/ko/topology/?p=capability%3Aagent-config-onboarding&mode=focus",
-      width: 1512,
-      height: 917,
-      markers: {
-        topologyRootAttentionWinner: "active-relation-inspector",
-        topologyAgentCurrentSurface: "selected-relation",
-        topologyAgentCurrentSurfaceRole: "active-relation-inspector",
-        topologyAgentCurrentSurfaceRoute:
-          "capability:agent-config-onboarding>element:operations-nav",
-        topologySelectedRelationCardHandoffContract:
-          "selected-relation-card-carries-mcp-cli-fallback",
-        topologySelectedRelationCardRoute: "source>target>type>action",
-        topologySelectedRelationCardEndpointRoute:
-          "capability:agent-config-onboarding>element:operations-nav",
-        topologySelectedRelationCardPrimaryAction: "explain_relation",
-        topologySelectedRelationCardCliFallback:
-          "ontology-atlas explain 'capability:agent-config-onboarding' 'element:operations-nav' [vault] --type 'contains'",
-        topologySelectedRelationCardSource: "capability:agent-config-onboarding",
-        topologySelectedRelationCardTarget: "element:operations-nav",
-        topologySelectedRelationCardType: "contains",
-        topologySelectedRelationCardLabelContextContract:
-          "selected-card-preserves-aggregate-label-context",
-        topologySelectedRelationCardLabelCount: 6,
-        topologySelectedRelationCardLabelVisibleText: "포함 ×6 · S1",
-        topologySelectedRelationCardLabelReadableText: "포함 ×6 · S1",
-        topologyFocusRelationLabelVisibleText: "포함 ×6 · S1",
-        topologySelectedRelationLabelCount: 6,
-        topologySelectedRelationClaimLensVisible: true,
-        topologySelectedRelationClaimLensText:
-          "강한 구조 · 출처 1개 · 타입이 있는 온톨로지 사실",
-        topologySelectedRelationClaimLensQuality: "strong",
-        topologySelectedRelationContractKind: "typed-fact-not-similarity",
-        topologySelectedRelationContractText:
-          "추론된 유사도 점수가 아니라 타입이 있는 온톨로지 사실입니다.",
-        topologySelectedRelationProofBandWidth: 236,
-        topologySelectedRelationProofBandHeight: 42,
-        topologySelectedRelationAgentDecisionText:
-          "타입과 근거가 있어 에이전트 전달에 포함할 수 있습니다.",
-        topologySelectedRelationEndpointRouteContract: "visible-source-target-names-wrap",
-        topologySelectedRelationEndpointRouteReadableText:
-          "Agent Config Onboarding → Operations Nav",
-        topologySelectedRelationCopyPayloadTool: "query_ontology",
-        topologySelectedRelationCopyPayloadAction: "explain_relation",
-        topologySelectedRelationCopyPayloadFrom: "capability:agent-config-onboarding",
-        topologySelectedRelationCopyPayloadTo: "element:operations-nav",
-        topologySelectedRelationCopyPayloadType: "contains",
-        topologySelectedRelationCopyPayloadEvidence: "source-backed",
-        topologySelectedRelationCopyPayloadGate: "handoff-ready",
-        topologySelectedRelationCopyPayloadCall:
-          'query_ontology({"operation":"explain_relation","from":"capability:agent-config-onboarding","to":"element:operations-nav","direction":"undirected","maxHops":5,"limit":10})',
-        topologySelectedRelationCopyPayloadSummary:
-          "query_ontology · explain_relation · capability:agent-config-onboarding → element:operations-nav · contains · source-backed · handoff-ready",
-        topologySelectedRelationCopyPayloadVisibleSummary: "설명 준비",
-        topologySelectedRelationCopyPayloadVisibleHandleSummary:
-          "agent-config-onboarding → operations-nav",
-        topologySelectedRelationCopyPayloadLayoutContract:
-          "visible-summary-and-handle-readable",
-      },
-    },
-    { capturedAt: "2026-06-22T12:00:00.000Z" },
-  );
-
-  assert.deepEqual(evidence.selectedRelationVisibleFactRouteProof, {
-    proof: "topology-selected-relation-visible-fact-route",
-    status: "proved",
-    route: "/ko/topology/?p=capability%3Aagent-config-onboarding&mode=focus",
-    root: {
-      attentionWinner: "active-relation-inspector",
-      currentSurface: "selected-relation",
-      currentSurfaceRole: "active-relation-inspector",
-      currentSurfaceRoute: "capability:agent-config-onboarding>element:operations-nav",
-    },
-    card: {
-      contract: "selected-relation-card-carries-mcp-cli-fallback",
-      route: "source>target>type>action",
-      endpointRoute: "capability:agent-config-onboarding>element:operations-nav",
-      primaryAction: "explain_relation",
-      cliFallback:
-        "ontology-atlas explain 'capability:agent-config-onboarding' 'element:operations-nav' [vault] --type 'contains'",
-      source: "capability:agent-config-onboarding",
-      target: "element:operations-nav",
-      type: "contains",
-    },
-    visibleFactRoute: {
-      claimLensVisible: true,
-      claimLensText: "강한 구조 · 출처 1개 · 타입이 있는 온톨로지 사실",
-      claimLensQuality: "strong",
-      contractKind: "typed-fact-not-similarity",
-      contractText: "추론된 유사도 점수가 아니라 타입이 있는 온톨로지 사실입니다.",
-      proofBandWidth: 236,
-      proofBandHeight: 42,
-      agentDecisionText: "타입과 근거가 있어 에이전트 전달에 포함할 수 있습니다.",
-      agentGateKind: "handoff-ready",
-      endpointRouteContract: "visible-source-target-names-wrap",
-      readableRoute: "Agent Config Onboarding → Operations Nav",
-    },
-    copyPayload: {
-      tool: "query_ontology",
-      action: "explain_relation",
-      from: "capability:agent-config-onboarding",
-      to: "element:operations-nav",
-      type: "contains",
-      evidence: "source-backed",
-      gate: "handoff-ready",
-      call:
-        'query_ontology({"operation":"explain_relation","from":"capability:agent-config-onboarding","to":"element:operations-nav","direction":"undirected","maxHops":5,"limit":10})',
-      summary:
-        "query_ontology · explain_relation · capability:agent-config-onboarding → element:operations-nav · contains · source-backed · handoff-ready",
-      visibleSummary: "설명 준비",
-      visibleHandleSummary: "agent-config-onboarding → operations-nav",
-      layoutContract: "visible-summary-and-handle-readable",
-    },
-    agentNextAction: "run-selected-relation-copy-payload",
-  });
-});
-
-test("WebView evidence flags slow connector label pass budget regressions", () => {
-  const evidence = verifier.buildWebviewEvidencePayload(
-    {
-      href: "tauri://localhost/ko/topology/?p=domain%3Aviews&mode=focus",
-      markers: {
-        topologyRepositionPassConnectorLabelMs: 3.2,
-        topologyRepositionMaxPassConnectorLabelMs: 3.2,
-        topologyRepositionPassSlowest: "connector-label",
-      },
-    },
-    { capturedAt: "2026-06-17T12:00:00.000Z" },
-  );
-
-  assert.deepEqual(evidence.connectorLabelPassProof, {
-    proof: "topology-connector-label-pass-budget",
-    status: "incomplete",
-    route: "/ko/topology/?p=domain%3Aviews&mode=focus",
-    passMs: 3.2,
-    budgetMs: 3,
-    maxPassMs: 3.2,
-    slowestPass: "connector-label",
-    agentNextAction: "read-relation-labels-after-connector-label-pass-budget",
-  });
-});
-
-test("WebView evidence summarizes selected focus dim context proof for agent handoff", () => {
-  const evidence = verifier.buildWebviewEvidencePayload(
-    {
-      href: "tauri://localhost/en/topology/?p=domain%3Aviews&mode=focus",
-      markers: {
-        topologyAttentionWinner: "focus-state",
-        topologyRootAttentionWinner: "focus-state",
-        topologyAgentCurrentSurface: "selected-node",
-        topologyAgentCurrentSurfaceRole: "active-node-inspector",
-        topologyAgentCurrentSurfaceRoute: "domain:views",
-        topologyUiScale: 1,
-        topologyUiScaleWritePolicy: "reuse-stable-scale",
-        topologyRootSelectedNodeId: "domain:views",
-        topologyNodePopoverVisible: true,
-        topologyNodePopoverSurfaceRole: "active-node-inspector",
-        topologyNodePopoverCollapsed: true,
-        topologySelectedNodePopoverVisible: true,
-        topologySelectedNodeId: "domain:views",
-        topologySelectedNodeTitle: "Views",
-        topologyNodePopoverRelationFactCount: "3",
-        topologyNodePopoverRelationTypeCount: "2",
-        topologyNodePopoverCompactMeaningText: "A core hub — 10 places depend on it",
-        topologyNodePopoverCompactMeaningContract: "plain-language-meaning-before-typed-facts",
-        topologyNodePopoverCompactMeaningResponsiveContract: "visible-desktop-sr-only-compact",
-        topologyNodePopoverCompactCommandRowContract: "facts-and-actions-share-final-scanline",
-        topologyNodePopoverCompactCommandRowGapToken:
-          "--topology-node-popover-compact-command-row-gap",
-        topologyNodePopoverCompactActionsContract: "actions-share-command-row-with-facts",
-        topologyNodePopoverCompactActionsReadableFlow:
-          "selected-node-facts-to-agent-handoff",
-        topologyNodePopoverCompactBriefActionReadableFlow:
-          "selected-node-facts-to-agent-brief",
-        topologyNodePopoverAgentHandoffContract: "selected-node-actions-visible",
-        topologyNodePopoverAgentHandoffRoute: "selected-node>facts>actions",
-        topologyNodePopoverAgentHandoffPrimaryAction: "focus-brief",
-        topologyNodePopoverAgentHandoffActionCount: "3",
-        topologyNodePopoverAgentHandoffRelationFactCount: "3",
-        topologyNodePopoverAgentHandoffRelationTypeCount: "2",
-        topologyNodePopoverAgentHandoffSummaryContract: "visible-mcp-cli-focus-brief",
-        topologyNodePopoverAgentHandoffVisibleSummary: "MCP/CLI · Brief",
-        topologyNodePopoverAgentHandoffSelectedNode: "domain:views",
-        topologyNodePopoverCompactRelationFactsContract:
-          "collapsed-dock-surfaces-typed-facts",
-        topologyNodePopoverCompactRelationFactsReadableContract:
-          "direct-typed-facts-not-scores",
-        topologyNodePopoverCompactRelationFactsNoScores:
-          "Typed ontology facts, not inferred similarity scores.",
-        topologyNodePopoverCompactRelationFactsAccessibleName:
-          "3 direct facts · 2 relation types · Typed ontology facts, not inferred similarity scores.",
-        topologyNodePopoverCompactRelationFactsTitle:
-          "3 direct facts · 2 relation types · Typed ontology facts, not inferred similarity scores.",
-        topologyNodePopoverCompactRelationFactsHandoffContract:
-          "compact-counts-route-to-relation-list-handoff",
-        topologyNodePopoverCompactRelationFactsHandoffRoute:
-          "selected-node>relations>fact>evidence>gate>action>payload",
-        topologyNodePopoverCompactRelationFactsHandoffTool: "query_ontology",
-        topologyNodePopoverCompactRelationFactsHandoffSummary:
-          "query_ontology · 2 rendered · 0 hidden · 3 direct facts",
-        topologyNodePopoverCompactRelationFactsHiddenRemainderCount: 0,
-        topologyNodePopoverCompactRelationFactsVisible: true,
-        topologyNodePopoverCompactActionsVisible: true,
-        topologyNodePopoverCompactRelationFactsTop: 212,
-        topologyNodePopoverCompactActionsTop: 206,
-        topologyNodePopoverCompactHandoffSummaryVisible: true,
-        topologyNodePopoverCompactHandoffSummaryContract: "visible-mcp-cli-focus-brief",
-        topologyNodePopoverCompactHandoffSummaryVisibleLabel: "MCP/CLI",
-        topologyNodePopoverCompactHandoffSummaryText: "MCP/CLI · Brief",
-        topologyNodePopoverCompactHandoffSummarySelectedNode: "domain:views",
-        topologyNodePopoverCompactHandoffSummaryClientWidth: 98,
-        topologyNodePopoverCompactHandoffSummaryScrollWidth: 98,
-        topologyNodePopoverCompactHandoffSummaryTop: 212,
-        topologyClickFocusRelationshipContext: "durable",
-        topologyClickFocusRelationshipContextSource: "selected-dock-companions",
-        topologyFocusClusterMode: "none",
-        topologyFocusClusterStage: "",
-        topologyFocusClusterSize: 6,
-        topologyFocusClusterVisible: false,
-        topologyDimOpacityContract: "readable-context-geography",
-        topologyDimAnchorOpacity: 0.26,
-        topologyDimChipOpacity: 0.08,
-        topologyDimContextOpacity: 0.08,
-        topologyDimAnchorVisibleCount: 3,
-        topologyDimChipVisibleCount: 4,
-        topologyDimAnchorMinOpacity: 0.26,
-        topologyDimChipMinOpacity: 0.08,
-      },
-    },
-    { capturedAt: "2026-06-20T12:00:00.000Z" },
-  );
-
-  assert.deepEqual(evidence.agentCurrentSurfaceProof, {
-    proof: "topology-agent-current-surface",
-    status: "proved",
-    route: "/en/topology/?p=domain%3Aviews&mode=focus",
-    attentionWinner: "focus-state",
-    currentSurface: "selected-node",
-    currentSurfaceRole: "active-node-inspector",
-    currentSurfaceRoute: "domain:views",
-    selectedNodeId: "domain:views",
-    rootSelectedNodeId: "domain:views",
-    handoff: {
-      contract: "selected-node-actions-visible",
-      route: "selected-node>facts>actions",
-      primaryAction: "focus-brief",
-      summaryContract: "visible-mcp-cli-focus-brief",
-      visibleSummary: "MCP/CLI · Brief",
-      actionCount: 3,
-      relationFactCount: 3,
-      relationTypeCount: 2,
-      agentNextAction: "copy-selected-node-focus-brief-or-expand-detail",
-    },
-    agentNextAction: "read-selected-node-surface-before-map-context",
-  });
-  assert.deepEqual(evidence.agentUiScaleStabilityProof, {
-    proof: "topology-ui-scale-stability",
-    status: "proved",
-    route: "/en/topology/?p=domain%3Aviews&mode=focus",
-    uiScale: 1,
-    writePolicy: "reuse-stable-scale",
-    stableScaleReused: true,
-    agentNextAction: "trust-stable-ui-scale-before-reading-surface-proof",
-  });
-  assert.deepEqual(evidence.selectedFocusDimProof, {
-    proof: "topology-selected-focus-dim-context",
-    status: "proved",
-    route: "/en/topology/?p=domain%3Aviews&mode=focus",
-    attention: {
-      winner: "focus-state",
-      selectedNodeId: "domain:views",
-      selectedNodeTitle: "Views",
-      compactMeaning: "A core hub — 10 places depend on it",
-      compactMeaningContract: "plain-language-meaning-before-typed-facts",
-      compactMeaningResponsiveContract: "visible-desktop-sr-only-compact",
-      positionContract: null,
-      gutterContract: null,
-      rightInsetToken: null,
-      rightInset: null,
-      supportContract: null,
-      commandRowContract: "facts-and-actions-share-final-scanline",
-      commandRowGapToken: "--topology-node-popover-compact-command-row-gap",
-      actionsContract: "actions-share-command-row-with-facts",
-      actionsReadableFlow: "selected-node-facts-to-agent-handoff",
-      relationFactsReadableContract: "direct-typed-facts-not-scores",
-      relationFactsAccessibleName:
-        "3 direct facts · 2 relation types · Typed ontology facts, not inferred similarity scores.",
-      relationFactsHandoff: {
-        contract: "compact-counts-route-to-relation-list-handoff",
-        route: "selected-node>relations>fact>evidence>gate>action>payload",
-        tool: "query_ontology",
-        summary: "query_ontology · 2 rendered · 0 hidden · 3 direct facts",
-        hiddenRemainderCount: 0,
-      },
-      briefActionReadableFlow: "selected-node-facts-to-agent-brief",
-      factsAndActionsShareScanline: true,
-      relationshipContext: "durable",
-      relationshipContextSource: "selected-dock-companions",
-      focusClusterMode: "none",
-      focusClusterStage: "",
-      focusClusterSize: 6,
-      focusClusterVisible: false,
-      hull: "not-rendered",
-    },
-    dim: {
-      contract: "readable-context-geography",
-      anchorOpacity: 0.26,
-      contextOpacity: 0.08,
-      contextOpacityAlias: 0.08,
-      anchorVisibleCount: 3,
-      contextVisibleCount: 4,
-      anchorMinOpacity: 0.26,
-      contextMinOpacity: 0.08,
-      anchorMinContract: 0.26,
-      contextMinContract: 0.08,
-      anchorToken: "--topology-map-dim-anchor-opacity",
-      contextToken: "--topology-map-dim-context-opacity",
-    },
-    agentNextAction: "read-selected-node-popover-before-background-map-context",
-  });
-  assert.deepEqual(evidence.nodePopoverCompactHandoffProof, {
-    proof: "topology-node-popover-compact-handoff-root",
-    status: "proved",
-    route: "/en/topology/?p=domain%3Aviews&mode=focus",
-    selectedNode: {
-      id: "domain:views",
-      title: "Views",
-      compactMeaning: "A core hub — 10 places depend on it",
-      relationFactCount: 3,
-      relationTypeCount: 2,
-    },
-    root: {
-      attentionWinner: "focus-state",
-      currentSurface: "selected-node",
-      currentSurfaceRole: "active-node-inspector",
-      currentSurfaceRoute: "domain:views",
-      selectedNodeId: "domain:views",
-    },
-    handoff: {
-      contract: "selected-node-actions-visible",
-      route: "selected-node>facts>actions",
-      primaryAction: "focus-brief",
-      summaryContract: "visible-mcp-cli-focus-brief",
-      visibleSummary: "MCP/CLI · Brief",
-      actionCount: 3,
-      relationFactCount: 3,
-      relationTypeCount: 2,
-      readableFlow: "selected-node-facts-to-agent-handoff",
-      briefActionFlow: "selected-node-facts-to-agent-brief",
-      compactSummary: {
-        visible: true,
-        contract: "visible-mcp-cli-focus-brief",
-        visibleLabel: "MCP/CLI",
-        text: "MCP/CLI · Brief",
-        selectedNode: "domain:views",
-        clientWidth: 98,
-        scrollWidth: 98,
-      },
-      relationFacts: {
-        contract: "compact-counts-route-to-relation-list-handoff",
-        route: "selected-node>relations>fact>evidence>gate>action>payload",
-        tool: "query_ontology",
-        summary: "query_ontology · 2 rendered · 0 hidden · 3 direct facts",
-        hiddenRemainderCount: 0,
-      },
-    },
-    agentNextAction: "copy-selected-node-focus-brief-or-expand-detail",
-  });
-});
-
-test("WebView evidence summarizes node popover compact verification proof for agent handoff", () => {
-  const evidence = verifier.buildWebviewEvidencePayload(
-    {
-      href: "tauri://localhost/en/topology/?p=domain%3Aviews&mode=focus",
-      markers: {
-        topologyNodePopoverVerifyAttempted: true,
-        topologyNodePopoverVerifyReason: "done",
-        topologyNodePopoverVerifyExpanded: true,
-        topologyNodePopoverVerifyCompactFactsVisible: true,
-        topologyNodePopoverVerifyCompactFactsContract: "collapsed-dock-surfaces-typed-facts",
-        topologyNodePopoverVerifyCompactFactsReadableContract: "direct-typed-facts-not-scores",
-        topologyNodePopoverVerifyCompactFactsAccessibleName:
-          "3 direct facts · 2 relation types · Typed ontology facts, not inferred similarity scores.",
-        topologyNodePopoverVerifyCompactFactsNoScores:
-          "Typed ontology facts, not inferred similarity scores.",
-        topologyNodePopoverVerifyCompactFactsHandoffContract:
-          "compact-counts-route-to-relation-list-handoff",
-        topologyNodePopoverVerifyCompactFactsHandoffRoute:
-          "selected-node>relations>fact>evidence>gate>action>payload",
-        topologyNodePopoverVerifyCompactFactsHandoffTool: "query_ontology",
-        topologyNodePopoverVerifyCompactFactsHandoffSummary:
-          "query_ontology · 2 rendered · 10 hidden · 15 direct facts",
-        topologyNodePopoverVerifyCompactFactsHiddenRemainderCount: 10,
-        topologyNodePopoverVerifyCompactActionsVisible: true,
-        topologyNodePopoverVerifyCompactActionsContract:
-          "actions-share-command-row-with-facts",
-        topologyNodePopoverVerifyCompactActionsReadableFlow:
-          "selected-node-facts-to-agent-handoff",
-        topologyNodePopoverVerifyCompactBriefVisible: true,
-        topologyNodePopoverVerifyCompactBriefAction: "copy-focus-brief",
-        topologyNodePopoverVerifyCompactBriefReadableFlow:
-          "selected-node-facts-to-agent-brief",
-        topologyNodePopoverVerifyCompactBriefRailLabel: "Agent handoff",
-        topologyNodePopoverVerifyCompactBriefTitle: "Agent handoff: Copy focus brief",
-      },
-    },
-    { capturedAt: "2026-06-20T12:10:00.000Z" },
-  );
-
-  assert.deepEqual(evidence.nodePopoverCompactVerificationProof, {
-    proof: "topology-node-popover-compact-verification",
-    status: "proved",
-    route: "/en/topology/?p=domain%3Aviews&mode=focus",
-    attempted: true,
-    expanded: true,
-    reason: "done",
-    compactFacts: {
-      visible: true,
-      contract: "collapsed-dock-surfaces-typed-facts",
-      readableContract: "direct-typed-facts-not-scores",
-      accessibleName:
-        "3 direct facts · 2 relation types · Typed ontology facts, not inferred similarity scores.",
-      noScores: "Typed ontology facts, not inferred similarity scores.",
-      handoffContract: "compact-counts-route-to-relation-list-handoff",
-      handoffRoute: "selected-node>relations>fact>evidence>gate>action>payload",
-      handoffTool: "query_ontology",
-      handoffSummary: "query_ontology · 2 rendered · 10 hidden · 15 direct facts",
-      hiddenRemainderCount: 10,
-    },
-    compactActions: {
-      visible: true,
-      contract: "actions-share-command-row-with-facts",
-      readableFlow: "selected-node-facts-to-agent-handoff",
-      briefVisible: true,
-      briefAction: "copy-focus-brief",
-      briefReadableFlow: "selected-node-facts-to-agent-brief",
-      briefRailLabel: "Agent handoff",
-      briefTitle: "Agent handoff: Copy focus brief",
-    },
-    agentNextAction: "read-compact-node-facts-before-expanded-popover-proof",
-  });
-});
-
-test("WebView evidence records unavailable visual evidence diagnostics for agent handoff", () => {
-  const evidence = verifier.buildWebviewEvidencePayload(
-    {
-      href: "tauri://localhost/ko/topology/?p=domain%3Aviews&mode=focus",
-      markers: {
-        topologyCreateNodeOpen: true,
-      },
-    },
-    {
-      capturedAt: "2026-06-16T12:05:00.000Z",
-      visualEvidence: {
-        screenshotPath: ".tmp/ontology-atlas-composer-blocking-ko.png",
-        screenshotStatus: "unavailable",
-        blocker: "screen-capture-returned-blank-image",
-        diagnosticsPath: ".tmp/ontology-atlas-composer-blocking-ko.png.diagnostics.json",
-        summary: "screencapture returned a blank or low-contrast image.",
-      },
-    },
-  );
-
-  assert.deepEqual(evidence.composerBlockingProof.agentHandoff.visualEvidence, {
-    screenshotPath: path.resolve(".tmp/ontology-atlas-composer-blocking-ko.png"),
-    screenshotStatus: "unavailable",
-    blocker: "screen-capture-returned-blank-image",
-    diagnosticsPath: path.resolve(".tmp/ontology-atlas-composer-blocking-ko.png.diagnostics.json"),
-    summary: "screencapture returned a blank or low-contrast image.",
-  });
-});
-
 test("WebView verification rejects stacked Relief transient surfaces", () => {
   const payload = {
     href: "tauri://localhost/ko/topology/",
@@ -2106,343 +764,6 @@ test("WebView verification rejects stacked Relief transient surfaces", () => {
   assert.match(
     validateWebviewVerifyPayload(payload, { expectedPath: "/ko/topology/" }),
     /stacked Relief transient surface contract/,
-  );
-});
-
-test("verify app launch args keep executable launch defaults", () => {
-  assert.deepEqual(
-    parseVerifyAppLaunchArgs([], {
-      defaultAppPath: "/tmp/Ontology Atlas.app",
-      defaultHoldMs: 5000,
-    }),
-    {
-      appPath: "/tmp/Ontology Atlas.app",
-      holdMs: 5000,
-      killExisting: false,
-      leaveRunning: false,
-      openApp: false,
-      requireWindow: false,
-      requireCapturableWindow: false,
-      requireAccessibilityWindow: false,
-      requireFrontmost: false,
-      requireWebviewContent: true,
-      requireWebviewRoute: null,
-      verifyTopologyDrag: false,
-      verifyTopologySelectedRelation: false,
-      verifyTopologyNodePopover: false,
-      verifyTopologyCreateNode: false,
-      verifyTopologyFocusNoop: false,
-      verifyTopologyFocusZoom: false,
-      verifyTopologyFrameProfile: false,
-      printWindowDiagnostics: false,
-      requireOwnerName: null,
-      minWindowSize: null,
-      minWebviewSize: null,
-      maxWebviewSize: null,
-      webviewWindowSize: null,
-      windowScreenshotPath: null,
-      tryWindowScreenshotPath: null,
-      webviewEvidencePath: null,
-      requireAccessibilityText: [],
-    },
-  );
-});
-
-test("verify app launch args keep LaunchServices dogfood compatible with window checks", () => {
-  assert.deepEqual(
-    parseVerifyAppLaunchArgs([
-      "/tmp/Custom.app",
-      "--open-app",
-      "--require-window",
-      "--require-capturable-window",
-      "--require-accessibility-window",
-    ]),
-    {
-      appPath: "/tmp/Custom.app",
-      holdMs: 5000,
-      killExisting: false,
-      leaveRunning: false,
-      openApp: true,
-      requireWindow: true,
-      requireCapturableWindow: true,
-      requireAccessibilityWindow: true,
-      requireFrontmost: false,
-      requireWebviewContent: false,
-      requireWebviewRoute: null,
-      verifyTopologyDrag: false,
-      verifyTopologySelectedRelation: false,
-      verifyTopologyNodePopover: false,
-      verifyTopologyCreateNode: false,
-      verifyTopologyFocusNoop: false,
-      verifyTopologyFocusZoom: false,
-      verifyTopologyFrameProfile: false,
-      printWindowDiagnostics: false,
-      requireOwnerName: null,
-      minWindowSize: null,
-      minWebviewSize: null,
-      maxWebviewSize: null,
-      webviewWindowSize: null,
-      windowScreenshotPath: null,
-      tryWindowScreenshotPath: null,
-      webviewEvidencePath: null,
-      requireAccessibilityText: [],
-    },
-  );
-});
-
-test("verify app launch args support stale-process cleanup, LaunchServices, and window checks", () => {
-  assert.deepEqual(
-    parseVerifyAppLaunchArgs([
-      "/tmp/Custom.app",
-      "--hold-ms=7000",
-      "--kill-existing",
-      "--leave-running",
-      "--open-app",
-      "--require-window",
-      "--require-capturable-window",
-      "--require-accessibility-window",
-      "--require-frontmost",
-      "--require-webview-content",
-      "--require-webview-route=/en/topology/",
-      "--verify-topology-drag",
-      "--verify-topology-create-node",
-      "--verify-topology-focus-noop",
-      "--print-window-diagnostics",
-      "--require-owner-name=Ontology Atlas",
-      "--min-window-size=1040x720",
-      "--min-webview-size=1400x860",
-      "--max-webview-size=1600x1000",
-      "--webview-window-size=1500x940",
-      "--window-screenshot=/tmp/ontology-atlas-window.png",
-      "--try-window-screenshot=/tmp/ontology-atlas-best-effort.png",
-      "--webview-evidence=/tmp/ontology-atlas-webview.json",
-      "--require-accessibility-text=개념 지도",
-      "--require-accessibility-text=AI 에이전트 그래프 검증",
-    ]),
-    {
-      appPath: "/tmp/Custom.app",
-      holdMs: 7000,
-      killExisting: true,
-      leaveRunning: true,
-      openApp: true,
-      requireWindow: true,
-      requireCapturableWindow: true,
-      requireAccessibilityWindow: true,
-      requireFrontmost: true,
-      requireWebviewContent: true,
-      requireWebviewRoute: "/en/topology/",
-      verifyTopologyDrag: true,
-      verifyTopologySelectedRelation: false,
-      verifyTopologyNodePopover: false,
-      verifyTopologyCreateNode: true,
-      verifyTopologyFocusNoop: true,
-      verifyTopologyFocusZoom: false,
-      verifyTopologyFrameProfile: false,
-      printWindowDiagnostics: true,
-      requireOwnerName: "Ontology Atlas",
-      minWindowSize: { width: 1040, height: 720 },
-      minWebviewSize: { width: 1400, height: 860 },
-      maxWebviewSize: { width: 1600, height: 1000 },
-      webviewWindowSize: { width: 1500, height: 940 },
-      windowScreenshotPath: "/tmp/ontology-atlas-window.png",
-      tryWindowScreenshotPath: "/tmp/ontology-atlas-best-effort.png",
-      webviewEvidencePath: "/tmp/ontology-atlas-webview.json",
-      requireAccessibilityText: ["개념 지도", "AI 에이전트 그래프 검증"],
-    },
-  );
-});
-
-test("verify app launch args normalize direct WebView route checks and allow route inspection", () => {
-  assert.equal(normalizeWebviewRoute("/en/topology/"), "/en/topology/");
-  assert.equal(normalizeWebviewRoute("/en/topology/?mode=path"), "/en/topology/?mode=path");
-  assert.equal(normalizeWebviewRoute(" /ko/ontology/ "), "/ko/ontology/");
-  assert.equal(normalizeWebviewRoute("en/topology/"), null);
-  assert.equal(normalizeWebviewRoute("//evil.test"), null);
-  assert.equal(normalizeWebviewRoute("https://evil.test/en/topology/"), null);
-  assert.equal(normalizeWebviewRoute("/en/topology/ bad"), null);
-  assert.deepEqual(
-    parseVerifyAppLaunchArgs([
-      "/tmp/Custom.app",
-      "--leave-running",
-      "--require-window",
-      "--require-webview-route=/en/topology/",
-    ]),
-    {
-      appPath: "/tmp/Custom.app",
-      holdMs: 5000,
-      killExisting: false,
-      leaveRunning: true,
-      openApp: false,
-      requireWindow: true,
-      requireCapturableWindow: false,
-      requireAccessibilityWindow: false,
-      requireFrontmost: false,
-      requireWebviewContent: true,
-      requireWebviewRoute: "/en/topology/",
-      verifyTopologyDrag: false,
-      verifyTopologySelectedRelation: false,
-      verifyTopologyNodePopover: false,
-      verifyTopologyCreateNode: false,
-      verifyTopologyFocusNoop: false,
-      verifyTopologyFocusZoom: false,
-      verifyTopologyFrameProfile: false,
-      printWindowDiagnostics: false,
-      requireOwnerName: null,
-      minWindowSize: null,
-      minWebviewSize: null,
-      maxWebviewSize: null,
-      webviewWindowSize: null,
-      windowScreenshotPath: null,
-      tryWindowScreenshotPath: null,
-      webviewEvidencePath: null,
-      requireAccessibilityText: [],
-    },
-  );
-});
-
-test("verify app launch waits until stale processes disappear after cleanup", async () => {
-  const calls = [];
-  const slept = [];
-  const remaining = await waitForExistingProcessesToExit({
-    appPath: "/tmp/Ontology Atlas.app",
-    executablePath: "/tmp/Ontology Atlas.app/Contents/MacOS/ontology-atlas",
-    timeoutMs: 1000,
-    intervalMs: 100,
-    readProcessIds: ({ appPath, executablePath }) => {
-      calls.push({ appPath, executablePath });
-      return calls.length < 3 ? [101, 202] : [];
-    },
-    sleepFn: async (ms) => {
-      slept.push(ms);
-    },
-  });
-
-  assert.deepEqual(remaining, []);
-  assert.equal(calls.length, 3);
-  assert.deepEqual(slept, [100, 100]);
-});
-
-test("verify app launch reports stale processes that survive cleanup polling", async () => {
-  const remaining = await waitForExistingProcessesToExit({
-    appPath: "/tmp/Ontology Atlas.app",
-    executablePath: "/tmp/Ontology Atlas.app/Contents/MacOS/ontology-atlas",
-    timeoutMs: 250,
-    intervalMs: 100,
-    readProcessIds: () => [303],
-    sleepFn: async () => undefined,
-  });
-
-  assert.deepEqual(remaining, [303]);
-});
-
-test("bundle path conflict warnings flag installed copies with the same bundle id", () => {
-  assert.deepEqual(
-    bundlePathConflictWarnings({
-      targetAppPath:
-        "/Users/me/ontology-atlas/src-tauri/target/release/bundle/macos/Ontology Atlas.app",
-      targetBundleIdentifier: "dev.jinan.ontology-atlas",
-      candidates: [
-        {
-          appPath: "/Applications/Ontology Atlas.app",
-          bundleIdentifier: "dev.jinan.ontology-atlas",
-        },
-        {
-          appPath: "/Users/me/Applications/Other.app",
-          bundleIdentifier: "com.example.other",
-        },
-        {
-          appPath:
-            "/Users/me/ontology-atlas/src-tauri/target/release/bundle/macos/Ontology Atlas.app",
-          bundleIdentifier: "dev.jinan.ontology-atlas",
-        },
-      ],
-    }),
-    [
-      "/Applications/Ontology Atlas.app shares bundle id dev.jinan.ontology-atlas with the verified app; app-name Computer Use may attach to that installed copy unless the Run script refreshed it, so use the full built app path when exact bundle provenance matters.",
-    ],
-  );
-});
-
-test("stale app cleanup prepares graceful quit commands before force killing", () => {
-  assert.deepEqual(
-    gracefulQuitExistingAppCommands({
-      appName: "Ontology Atlas",
-      bundleIdentifier: "dev.jinan.ontology-atlas",
-    }),
-    [
-      {
-        command: "osascript",
-        args: ["-e", 'tell application id "dev.jinan.ontology-atlas" to quit'],
-      },
-      {
-        command: "osascript",
-        args: ["-e", 'tell application "Ontology Atlas" to quit'],
-      },
-    ],
-  );
-});
-
-test("graceful quit commands time out so verification can fall back to process termination", () => {
-  assert.deepEqual(gracefulQuitCommandOptions(), {
-    stdio: "ignore",
-    timeout: 1200,
-  });
-});
-
-test("Accessibility text probe script targets launched pids", () => {
-  const script = buildAccessibilityTextProbeSwift([101, 202], ["개념 지도"]);
-
-  assert.match(script, /let requiredPids: Set<pid_t> = \[101,202\]/);
-  assert.match(script, /let requiredText = \["개념 지도"\]/);
-  assert.match(script, /func isComplete/);
-  assert.match(script, /func collectText/);
-  assert.match(script, /kAXChildrenAttribute/);
-});
-
-test("validateAccessibilityText requires every requested text fragment", () => {
-  const payload = "Ontology Atlas\n개념 지도\nAI 에이전트 그래프 검증";
-
-  assert.equal(validateAccessibilityText(payload, []), null);
-  assert.equal(
-    validateAccessibilityText(payload, ["개념 지도", "AI 에이전트 그래프 검증"]),
-    null,
-  );
-  assert.match(
-    validateAccessibilityText(payload, ["Source Vault"]),
-    /missing Accessibility text "Source Vault"/,
-  );
-  assert.match(validateAccessibilityText("", ["개념 지도"]), /empty Accessibility text/);
-});
-
-test("verify app launch lock prevents concurrent app checks and releases cleanly", () => {
-  const tmp = fs.mkdtempSync(path.join(os.tmpdir(), "ontology-atlas-lock-test-"));
-  const lockDir = path.join(tmp, "verify.lock");
-  try {
-    const first = createVerifyLock(lockDir, { appPath: "/tmp/Ontology Atlas.app" });
-    assert.equal(first.ok, true);
-
-    const second = createVerifyLock(lockDir, { appPath: "/tmp/Ontology Atlas.app" });
-    assert.equal(second.ok, false);
-    assert.match(second.message, /another desktop app verification is already running/);
-
-    first.release();
-    const third = createVerifyLock(lockDir, { appPath: "/tmp/Ontology Atlas.app" });
-    assert.equal(third.ok, true);
-    third.release();
-    assert.equal(fs.existsSync(lockDir), false);
-  } finally {
-    fs.rmSync(tmp, { recursive: true, force: true });
-  }
-});
-
-test("verify app launch lock path is stable per app path", () => {
-  assert.equal(
-    verifyLockPath("/Applications/Ontology Atlas.app"),
-    verifyLockPath("/Applications/Ontology Atlas.app"),
-  );
-  assert.notEqual(
-    verifyLockPath("/Applications/Ontology Atlas.app"),
-    verifyLockPath("/tmp/Ontology Atlas.app"),
   );
 });
 
@@ -11437,1326 +9758,5 @@ test("WebView verification payload parses nested JSON and checks loaded DOM", ()
       },
     }, { expectedPath: "/en/topology/" }),
     /malformed Relief relation lens copy/,
-  );
-});
-
-test("WebView verification payload parser uses the latest reported DOM snapshot", () => {
-  const loadingPayload = {
-    href: "tauri://localhost/en/",
-    title: "Ontology Atlas",
-    bodyText: "Loading local app shell",
-    bodyChildren: 1,
-    readyState: "loading",
-    markers: {},
-    width: 1,
-    height: 1,
-  };
-  const loadedPayload = {
-    ...loadingPayload,
-    bodyText: "Workspace\nOntology\nRelief\nConcept map",
-    bodyChildren: 19,
-    readyState: "complete",
-    markers: {
-      ontologyNav: true,
-      sourceVaultNav: true,
-      agentBriefCopy: false,
-      businessDecisionQuestions: false,
-      readerDecisionLens: false,
-      topologyRelief: false,
-      topologyCardCount: 0,
-      topologyCardOverlapCount: 0,
-      topologyCardClippedCount: 0,
-      topologyFixedSurfaceCount: 0,
-          topologyFixedSurfaceMeasureContract: "single-pass-rect-read",
-      topologyCardFixedSurfaceOverlapCount: 0,
-      topologyDragAttempted: false,
-      topologyDragReason: "",
-      topologyDragFocusMoved: false,
-      topologyDragFocusDelta: null,
-      topologyDragCompanionVisible: false,
-      topologyDragCompanionAligned: false,
-      topologyDragCompanionDelta: null,
-      topologyDragCompanionSlug: "",
-      topologyDragCompanionCount: 0,
-      topologyDragVisibleCompanionCount: 0,
-      topologyDragAlignedCompanionCount: 0,
-    },
-    width: 1280,
-    height: 789,
-  };
-  const stdout = [
-    `[ontology-atlas-webview-verify] ${JSON.stringify(JSON.stringify(loadingPayload))}`,
-    `[ontology-atlas-webview-verify] ${JSON.stringify(JSON.stringify(loadedPayload))}`,
-  ].join("\n");
-
-  assert.deepEqual(parseWebviewVerifyPayload(stdout), loadedPayload);
-  assert.equal(validateWebviewVerifyPayload(parseWebviewVerifyPayload(stdout)), null);
-});
-
-test("WebView verification payload parser skips empty in-flight snapshots", () => {
-  const payload = {
-    href: "tauri://localhost/ko/topology/?mode=path",
-    title: "지형도 · ontology-atlas",
-    bodyText: "온톨로지 지형도\n후보 11/21개 표시 중",
-    bodyChildren: 19,
-    readyState: "complete",
-    markers: {
-      topologyRelief: true,
-    },
-    width: 1100,
-    height: 768,
-  };
-  const stdout = [
-    `[ontology-atlas-webview-verify] ${JSON.stringify(JSON.stringify(payload))}`,
-    "[ontology-atlas-webview-verify] ",
-  ].join("\n");
-
-  assert.deepEqual(parseWebviewVerifyPayload(stdout), payload);
-});
-
-test("selected relation label compact markers match rendered width and viewport bounds", () => {
-  const baseMarkers = {
-    topologySelectedRelationLabelHitWidth: 144,
-    topologySelectedRelationLabelHitHeight: 32,
-    topologySelectedRelationLabelHitLeft: 320,
-    topologySelectedRelationLabelHitRight: 464,
-    topologySelectedRelationLabelCompact: "false",
-    topologySelectedRelationLabelDesiredWidth: 144,
-    topologySelectedRelationLabelCenteredAvailableWidth: 144,
-    topologySelectedRelationLabelViewportClampContract: "centered-within-viewport",
-    topologySelectedRelationLabelViewportClampSide: "none",
-    topologySelectedRelationLabelViewportInset: 16,
-  };
-
-  assert.equal(validateSelectedRelationLabelCompactMarkers(baseMarkers, 1512), null);
-  assert.equal(
-    validateSelectedRelationLabelCompactMarkers(
-      {
-        ...baseMarkers,
-        topologySelectedRelationLabelHitWidth: 150,
-        topologySelectedRelationLabelHitLeft: 16,
-        topologySelectedRelationLabelHitRight: 166,
-        topologySelectedRelationLabelCompact: "true",
-        topologySelectedRelationLabelDesiredWidth: 314,
-        topologySelectedRelationLabelCenteredAvailableWidth: 150,
-        topologySelectedRelationLabelViewportClampContract: "compacted-to-viewport-edge",
-        topologySelectedRelationLabelViewportClampSide: "left",
-      },
-      1512,
-    ),
-    null,
-  );
-  assert.match(
-    validateSelectedRelationLabelCompactMarkers(
-      {
-        ...baseMarkers,
-        topologySelectedRelationLabelHitWidth: 150,
-        topologySelectedRelationLabelHitLeft: 16,
-        topologySelectedRelationLabelHitRight: 166,
-        topologySelectedRelationLabelCompact: "false",
-        topologySelectedRelationLabelDesiredWidth: 314,
-        topologySelectedRelationLabelCenteredAvailableWidth: 150,
-        topologySelectedRelationLabelViewportClampContract: "compacted-to-viewport-edge",
-        topologySelectedRelationLabelViewportClampSide: "left",
-      },
-      1512,
-    ),
-    /compact marker was inconsistent/,
-  );
-  assert.match(
-    validateSelectedRelationLabelCompactMarkers(
-      {
-        ...baseMarkers,
-        topologySelectedRelationLabelHitLeft: 8,
-        topologySelectedRelationLabelHitRight: 322,
-      },
-      1512,
-    ),
-    /overflowed the viewport left/,
-  );
-  assert.match(
-    validateSelectedRelationLabelCompactMarkers(
-      {
-        ...baseMarkers,
-        topologySelectedRelationLabelViewportClampContract: "compacted-to-viewport-edge",
-        topologySelectedRelationLabelViewportClampSide: "none",
-      },
-      1512,
-    ),
-    /edge compaction without a clamp side/,
-  );
-});
-
-test("relation label frame geometry markers require after-render ready counts", () => {
-  const markers = {
-    topologyRelationLabelGeometryContract: "frame-positioned-hit-targets",
-    topologyRelationLabelGeometrySource: "after-render-layout-pass",
-    topologyRelationLabelGeometryExpectedCount: 2,
-    topologyRelationLabelGeometryReadyCount: 2,
-    topologyRelationLabelGeometryPendingCount: 0,
-  };
-
-  assert.equal(validateRelationLabelFrameGeometryMarkers(markers), null);
-  assert.match(
-    validateRelationLabelFrameGeometryMarkers({
-      ...markers,
-      topologyRelationLabelGeometrySource: "pending-frame",
-    }),
-    /frame geometry source/,
-  );
-  assert.match(
-    validateRelationLabelFrameGeometryMarkers({
-      ...markers,
-      topologyRelationLabelGeometryReadyCount: 1,
-    }),
-    /ready count/,
-  );
-  assert.match(
-    validateRelationLabelFrameGeometryMarkers({
-      ...markers,
-      topologyRelationLabelGeometryPendingCount: 1,
-    }),
-    /pending labels/,
-  );
-});
-
-test("topology connector cache markers require frame fallback with zero DOM reads", () => {
-  const markers = {
-    topologyConnectorDomIndexContract: "reuse-card-index",
-    topologyConnectorRectCacheContract: "frame-local-card-rect-cache",
-    topologyConnectorRectCacheFrameFallbackContract:
-      "reuse-card-placement-frame-rects-before-dom-read",
-    topologyConnectorRectCacheAccounting: "reads-plus-hits",
-    topologyConnectorRectCacheSize: 7,
-    topologyConnectorRectCacheReadCount: 0,
-    topologyConnectorRectCacheHitCount: 26,
-  };
-
-  assert.equal(validateTopologyConnectorCacheMarkers(markers), null);
-  assert.match(
-    validateTopologyConnectorCacheMarkers({
-      ...markers,
-      topologyConnectorRectCacheFrameFallbackContract: "direct-dom-read",
-    }),
-    /frame fallback contract/,
-  );
-  assert.match(
-    validateTopologyConnectorCacheMarkers({
-      ...markers,
-      topologyConnectorRectCacheReadCount: 1,
-    }),
-    /connector rect cache proof/,
-  );
-  assert.match(
-    validateTopologyConnectorCacheMarkers({
-      ...markers,
-      topologyConnectorRectCacheHitCount: 0,
-    }),
-    /connector rect cache proof/,
-  );
-});
-
-test("selected relation card attention lane keeps map label and support panel clear", () => {
-  const baseMarkers = {
-    topologySelectedRelationCardDockContract: "right-compact-relation-rail",
-    topologySelectedRelationCardAttentionLane: "right-inspector-rail",
-    topologySelectedRelationCardMapClearanceContract: "selected-label-keeps-map-lane",
-    topologySelectedRelationCardLeft: 1209,
-    topologySelectedRelationCardRight: 1476,
-    topologySelectedRelationLabelHitRight: 693,
-    topologyAnalysisPanelVisible: false,
-    topologyAnalysisPanelRight: 0,
-  };
-
-  assert.equal(validateSelectedRelationCardAttentionLane(baseMarkers, 1512), null);
-  assert.match(
-    validateSelectedRelationCardAttentionLane(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCardRight: 1502,
-      },
-      1512,
-    ),
-    /left the right inspector lane/,
-  );
-  assert.match(
-    validateSelectedRelationCardAttentionLane(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCardLeft: 704,
-      },
-      1512,
-    ),
-    /crowded the selected relation label/,
-  );
-  assert.match(
-    validateSelectedRelationCardAttentionLane(
-      {
-        ...baseMarkers,
-        topologyAnalysisPanelVisible: true,
-        topologyAnalysisPanelRight: 1190,
-      },
-      1512,
-    ),
-    /crowded the support panel/,
-  );
-});
-
-test("selected relation endpoint route markers prove visible source and target names", () => {
-  const baseMarkers = {
-    topologySelectedRelationCopyPayloadFrom: "domain:views",
-    topologySelectedRelationCopyPayloadTo: "capability:topology-analysis-modes",
-    topologySelectedRelationHandleStripSource: "domain:views",
-    topologySelectedRelationHandleStripTarget: "capability:topology-analysis-modes",
-    topologySelectedRelationEndpointRouteContract: "visible-source-target-names-wrap",
-    topologySelectedRelationEndpointRouteSourceName: "Views",
-    topologySelectedRelationEndpointRouteTargetName: "Topology Analysis Modes",
-    topologySelectedRelationEndpointRouteWrapPolicy:
-      "wrap-allowed-no-horizontal-overflow",
-    topologySelectedRelationEndpointRouteLineBudget: "2",
-    topologySelectedRelationEndpointRouteSourceHandle: "domain:views",
-    topologySelectedRelationEndpointRouteTargetHandle: "capability:topology-analysis-modes",
-    topologySelectedRelationEndpointRouteHandleSummary:
-      "domain:views → capability:topology-analysis-modes",
-    topologySelectedRelationEndpointRouteReadableText:
-      "Views → Topology Analysis Modes",
-    topologySelectedRelationEndpointReadableRoute: "Views → Topology Analysis Modes",
-    topologySelectedRelationEndpointRouteText: "Views→Topology Analysis Modes",
-    topologySelectedRelationEndpointRouteWidth: 212,
-    topologySelectedRelationEndpointRouteHeight: 30,
-    topologySelectedRelationEndpointRouteClientWidth: 212,
-    topologySelectedRelationEndpointRouteScrollWidth: 212,
-  };
-
-  assert.equal(validateSelectedRelationEndpointRouteMarkers(baseMarkers), null);
-  assert.match(
-    validateSelectedRelationEndpointRouteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEndpointRouteContract: "hidden-handles-only",
-    }),
-    /endpoint route contract/,
-  );
-  assert.match(
-    validateSelectedRelationEndpointRouteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEndpointRouteWrapPolicy: "truncate",
-    }),
-    /endpoint route wrap policy/,
-  );
-  assert.match(
-    validateSelectedRelationEndpointRouteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEndpointRouteLineBudget: "1",
-    }),
-    /endpoint route line budget/,
-  );
-  assert.match(
-    validateSelectedRelationEndpointRouteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEndpointRouteText: "Views",
-    }),
-    /endpoint names not visible/,
-  );
-  assert.match(
-    validateSelectedRelationEndpointRouteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEndpointRouteReadableText: "Views",
-    }),
-    /endpoint readable route/,
-  );
-  assert.match(
-    validateSelectedRelationEndpointRouteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEndpointReadableRoute: "Views",
-    }),
-    /endpoint layer readable route/,
-  );
-  assert.match(
-    validateSelectedRelationEndpointRouteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEndpointRouteTargetHandle: "capability:wrong",
-    }),
-    /visible endpoint handles/,
-  );
-  assert.match(
-    validateSelectedRelationEndpointRouteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEndpointRouteClientWidth: 180,
-      topologySelectedRelationEndpointRouteScrollWidth: 212,
-    }),
-    /overflowing Relief selected relation endpoint route/,
-  );
-});
-
-test("selected relation identity markers expose the root selected edge id", () => {
-  const baseMarkers = {
-    topologySelectedRelationVerifySelected: true,
-    topologyAgentCurrentSurface: "selected-relation",
-    topologyAgentCurrentSurfaceRoute: "domain:views>capability:topology-analysis-modes",
-    topologySelectedRelationEdgeId: "geid_138_17",
-    topologySelectedRelationLabelSource: "domain:views",
-    topologySelectedRelationLabelTarget: "capability:topology-analysis-modes",
-    topologySelectedRelationLabelRoute: "domain:views>capability:topology-analysis-modes",
-  };
-
-  assert.equal(validateSelectedRelationIdentityMarkers(baseMarkers), null);
-  assert.match(
-    validateSelectedRelationIdentityMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEdgeId: "",
-    }),
-    /selected relation edge id/,
-  );
-  assert.match(
-    validateSelectedRelationIdentityMarkers({
-      ...baseMarkers,
-      topologyAgentCurrentSurfaceRoute: "domain:wrong>capability:topology-analysis-modes",
-    }),
-    /selected relation route/,
-  );
-});
-
-test("selected relation endpoint visibility markers prove source and target cards stay readable", () => {
-  const baseMarkers = {
-    topologyCardsReady: true,
-    topologySelectedRelationCardSource: "domain:views",
-    topologySelectedRelationCardTarget: "capability:topology-analysis-modes",
-    topologySelectedRelationEndpointVisibilityContract:
-      "selected-relation-keeps-source-target-readable",
-    topologySelectedRelationEndpointExpectedCount: 2,
-    topologySelectedRelationEndpointVisibleCount: 2,
-    topologySelectedRelationEndpointHiddenCount: 0,
-    topologySelectedRelationEndpointCards: [
-      {
-        slug: "domain:views",
-        roleBadgeText: "FROM",
-        roleBadgeContract: "visible-source-target-role-badge",
-        roleBadgeVisible: true,
-        visible: true,
-        surfaceHidden: "",
-        shift: "safe-shift",
-      },
-      {
-        slug: "capability:topology-analysis-modes",
-        roleBadgeText: "TO",
-        roleBadgeContract: "visible-source-target-role-badge",
-        roleBadgeVisible: true,
-        visible: true,
-        surfaceHidden: "",
-        shift: "safe-shift",
-      },
-    ],
-  };
-
-  assert.equal(validateSelectedRelationEndpointVisibilityMarkers(baseMarkers), null);
-  assert.match(
-    validateSelectedRelationEndpointVisibilityMarkers({
-      ...baseMarkers,
-      topologyCardsReady: false,
-    }),
-    /skeleton card layer was not ready/,
-  );
-  assert.match(
-    validateSelectedRelationEndpointVisibilityMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEndpointVisibilityContract: "best-effort",
-    }),
-    /endpoint visibility contract/,
-  );
-  assert.match(
-    validateSelectedRelationEndpointVisibilityMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEndpointHiddenCount: 1,
-    }),
-    /endpoint visibility proof/,
-  );
-  assert.match(
-    validateSelectedRelationEndpointVisibilityMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEndpointCards: [
-        {
-          slug: "domain:views",
-          visible: true,
-          surfaceHidden: "",
-          shift: "safe-shift",
-        },
-        {
-          slug: "capability:wrong",
-          visible: true,
-          surfaceHidden: "",
-          shift: "safe-shift",
-        },
-      ],
-    }),
-    /without source and target/,
-  );
-  assert.match(
-    validateSelectedRelationEndpointVisibilityMarkers({
-      ...baseMarkers,
-      topologySelectedRelationEndpointCards: [
-        {
-          slug: "domain:views",
-          visible: true,
-          surfaceHidden: "",
-          shift: "safe-shift",
-        },
-        {
-          slug: "capability:topology-analysis-modes",
-          visible: false,
-          surfaceHidden: "true",
-          shift: "safe-shift",
-        },
-      ],
-    }),
-    /hidden Relief selected relation endpoint card/,
-  );
-});
-
-test("selected relation context silhouette markers suppress lower-priority background cards", () => {
-  const baseMarkers = {
-    topologySelectedRelationContextSilhouettePolicy:
-      "selected-relation-keeps-endpoints-and-orientation-anchors-only",
-    topologySelectedRelationContextSilhouetteActive: true,
-    topologySelectedRelationContextSilhouetteHiddenCount: 6,
-    topologySelectedRelationLowerPriorityVisibleDimmedCount: 0,
-    topologySelectedRelationVisibleOrientationAnchorCount: 3,
-    topologySelectedRelationHiddenContextInteractionContract:
-      "hidden-context-is-not-pointer-focus-or-a11y-target",
-    topologySelectedRelationHiddenContextInteractiveCount: 0,
-  };
-
-  assert.equal(validateSelectedRelationContextSilhouetteMarkers(baseMarkers), null);
-  assert.match(
-    validateSelectedRelationContextSilhouetteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationContextSilhouettePolicy: "all-context-visible",
-    }),
-    /context silhouette policy/,
-  );
-  assert.match(
-    validateSelectedRelationContextSilhouetteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationContextSilhouetteActive: false,
-    }),
-    /did not activate/,
-  );
-  assert.match(
-    validateSelectedRelationContextSilhouetteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationLowerPriorityVisibleDimmedCount: 2,
-    }),
-    /noisy Relief selected relation context/,
-  );
-  assert.match(
-    validateSelectedRelationContextSilhouetteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationHiddenContextInteractionContract:
-        "hidden-context-is-only-transparent",
-    }),
-    /hidden context interaction contract/,
-  );
-  assert.match(
-    validateSelectedRelationContextSilhouetteMarkers({
-      ...baseMarkers,
-      topologySelectedRelationHiddenContextInteractiveCount: 1,
-    }),
-    /interactive hidden Relief selected relation context/,
-  );
-});
-
-test("selected relation card density contract keeps the relation inspector compact", () => {
-  const baseMarkers = {
-    topologySelectedRelationCardDensity: "compact",
-    topologySelectedRelationCardDensityContract: "mini-relation-inspector",
-    topologySelectedRelationCardScaleContract: "density-fixed-no-ui-zoom",
-    topologySelectedRelationCardTypographyContract: "legible-compact-relation-inspector",
-    topologySelectedRelationCardMaxHeightToken: "--topology-selected-relation-card-max-height",
-    topologySelectedRelationCardKickerFontSizeToken:
-      "--topology-selected-relation-kicker-font-size",
-    topologySelectedRelationCardChipFontSizeToken:
-      "--topology-selected-relation-chip-font-size",
-    topologySelectedRelationCardRouteLabelFontSizeToken:
-      "--topology-selected-relation-route-label-font-size",
-    topologySelectedRelationCardRouteValueFontSizeToken:
-      "--topology-selected-relation-route-value-font-size",
-    topologySelectedRelationCardPayloadFontSizeToken:
-      "--topology-selected-relation-payload-font-size",
-    topologySelectedRelationAgentRouteSteps: [
-      { kind: "fact", labelFontSize: "8px", valueFontSize: "10px" },
-      { kind: "evidence", labelFontSize: "8px", valueFontSize: "10px" },
-      { kind: "action", visibility: "metadata-only" },
-    ],
-    topologySelectedRelationCardWidth: 248,
-    topologySelectedRelationCardHeight: 222,
-    topologySelectedRelationProofBandHeight: 36,
-    topologySelectedRelationCopyActionRailHeight: 29,
-    topologySelectedRelationCopyPayloadHeight: 42,
-    topologySelectedRelationAgentRouteHeight: 29,
-    topologySelectedRelationPrimaryCopyActionKind: "explain_relation",
-    topologySelectedRelationCommandCueContract:
-      "visible-compact-primary-command-cue",
-    topologySelectedRelationCommandCueVisibleText: "Next step · Explain",
-    topologySelectedRelationCommandCueAction: "explain_relation",
-  };
-
-  assert.equal(validateSelectedRelationCardDensityContract(baseMarkers, 1512), null);
-  assert.equal(validateSelectedRelationCardDensityContract(baseMarkers, 1920), null);
-  assert.equal(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCardWidth: 352,
-      },
-      2560,
-    ),
-    null,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCardWidth: 352,
-      },
-      1512,
-    ),
-    /oversized compact Relief selected relation card/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(baseMarkers, 2560),
-    /cramped wide Relief selected relation card/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCardScaleContract: "zoomed-with-topology-ui",
-      },
-      1920,
-    ),
-    /scale contract/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCardTypographyContract: "micro-type",
-      },
-      1512,
-    ),
-    /typography contract/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCardRouteValueFontSizeToken:
-          "--topology-selected-relation-old-route-value-font-size",
-      },
-      1512,
-    ),
-    /typography token/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationAgentRouteSteps: [
-          { kind: "fact", labelFontSize: "7px", valueFontSize: "9px" },
-        ],
-      },
-      1512,
-    ),
-    /too-small Relief selected relation route typography/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCardHeight: 336,
-      },
-      1512,
-    ),
-    /oversized compact Relief selected relation card/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationProofBandHeight: 52,
-      },
-      1512,
-    ),
-    /oversized Relief selected relation proof band/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCopyActionRailHeight: 44,
-      },
-      1512,
-    ),
-    /oversized Relief selected relation copy action rail/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCopyPayloadHeight: 56,
-      },
-      1512,
-    ),
-    /oversized Relief selected relation copy payload strip/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationAgentRouteHeight: 48,
-      },
-      1512,
-    ),
-    /oversized Relief selected relation agent route rail/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCommandCueContract: "sr-only-command",
-      },
-      1512,
-    ),
-    /command cue contract/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCommandCueVisibleText: "",
-      },
-      1512,
-    ),
-    /command cue text/,
-  );
-  assert.match(
-    validateSelectedRelationCardDensityContract(
-      {
-        ...baseMarkers,
-        topologySelectedRelationCommandCueAction: "relation_check",
-      },
-      1512,
-    ),
-    /command cue action/,
-  );
-});
-
-test("WebView verification waits for the latest snapshot that passes route gates", async () => {
-  const pendingPayload = {
-    href: "tauri://localhost/en/topology/",
-    title: "Relief · ontology-atlas",
-    bodyText:
-      "Workspace\nOntology\nRelief\n292 concepts\n21 concept cards\nShowing the readable card skeleton.",
-    bodyChildren: 19,
-    readyState: "complete",
-    markers: {
-      ontologyNav: true,
-      sourceVaultNav: true,
-      agentBriefCopy: false,
-      businessDecisionQuestions: false,
-      readerDecisionLens: false,
-      topologyRelief: true,
-      topologyStagePanClickCancelPx: 12,
-      topologyCardsReady: false,
-      topologyCardCount: 0,
-      topologyCardOverlapCount: 0,
-      topologyCardClippedCount: 0,
-      topologyFixedSurfaceCount: 2,
-          topologyFixedSurfaceMeasureContract: "single-pass-rect-read",
-      topologyFixedSurfaceOverlapCount: 0,
-      topologyFixedSurfaceOverlapSample: [],
-      topologyCardFixedSurfaceOverlapCount: 0,
-      topologyRelationLensVisible: false,
-      topologyRelationLensText: "",
-      topologyRelationLensPluralMismatch: false,
-      topologyRelationQualityLensVisible: false,
-      topologyRelationQualityLensText: "",
-      topologyOverviewAgentReadinessText: "",
-      topologyOverviewAgentReadinessMeterSegments: [],
-    },
-    width: 1280,
-    height: 789,
-  };
-  const readyPayload = {
-    ...pendingPayload,
-    markers: {
-      ...pendingPayload.markers,
-      topologyCardsReady: true,
-      topologyCardCount: 21,
-      topologyRelationLensVisible: true,
-      topologyRelationLensText:
-        "Relation lens · 21 direct facts · 1 relation type · Typed ontology facts, not inferred similarity scores.",
-      topologyRelationQualityLensVisible: true,
-      topologyRelationQualityLensText: "Relation quality: strong 1 · supported 1 · weak 0 · review 0",
-      topologyOverviewAgentReadinessText:
-        "Agent readiness: handoff-ready 2 · preflight 0 · review 0",
-      topologyOverviewAgentReadinessMeterSegments: [
-        { kind: "ready", count: "2" },
-        { kind: "preflight", count: "0" },
-        { kind: "review", count: "0" },
-      ],
-    },
-  };
-  let stdout = `[ontology-atlas-webview-verify] ${JSON.stringify(JSON.stringify(pendingPayload))}\n`;
-  const result = await waitForWebviewVerifyPayload(
-    () => stdout,
-    {
-      timeoutMs: 500,
-      intervalMs: 10,
-      validatePayload: (candidate) =>
-        validateWebviewVerifyPayload(candidate, { expectedPath: "/en/topology/" }),
-    },
-  );
-  assert.deepEqual(result.payload, pendingPayload);
-  assert.match(String(result.validationError), /skeleton overlay/);
-
-  stdout += `[ontology-atlas-webview-verify] ${JSON.stringify(JSON.stringify(readyPayload))}\n`;
-  const readyResult = await waitForWebviewVerifyPayload(
-    () => stdout,
-    {
-      timeoutMs: 500,
-      intervalMs: 10,
-      validatePayload: (candidate) =>
-        validateWebviewVerifyPayload(candidate, { expectedPath: "/en/topology/" }),
-    },
-  );
-  assert.deepEqual(readyResult, { payload: readyPayload, validationError: null });
-});
-
-test("parseMinWindowSize accepts WIDTHxHEIGHT only", () => {
-  assert.deepEqual(parseMinWindowSize("1280x820"), { width: 1280, height: 820 });
-  assert.equal(parseMinWindowSize("1280"), null);
-  assert.equal(parseMinWindowSize("widextall"), null);
-});
-
-test("parseOnscreenWindows keeps visible layer-zero windows for launched process ids", () => {
-  const payload = JSON.stringify([
-    {
-      kCGWindowOwnerPID: 101,
-      kCGWindowIsOnscreen: true,
-      kCGWindowLayer: 0,
-      kCGWindowAlpha: 1,
-      kCGWindowOwnerName: "Ontology Atlas",
-      kCGWindowBounds: { Width: 1280, Height: 820 },
-    },
-    {
-      kCGWindowOwnerPID: 101,
-      kCGWindowIsOnscreen: true,
-      kCGWindowLayer: 1,
-      kCGWindowAlpha: 1,
-      kCGWindowBounds: { Width: 1280, Height: 820 },
-    },
-    {
-      kCGWindowOwnerPID: 202,
-      kCGWindowIsOnscreen: true,
-      kCGWindowLayer: 0,
-      kCGWindowAlpha: 1,
-      kCGWindowBounds: { Width: 1280, Height: 820 },
-    },
-    {
-      kCGWindowOwnerPID: 101,
-      kCGWindowIsOnscreen: false,
-      kCGWindowLayer: 0,
-      kCGWindowAlpha: 1,
-      kCGWindowBounds: { Width: 1280, Height: 820 },
-    },
-  ]);
-
-  assert.deepEqual(parseOnscreenWindows(payload, [101]), [
-    {
-      kCGWindowOwnerPID: 101,
-      kCGWindowIsOnscreen: true,
-      kCGWindowLayer: 0,
-      kCGWindowAlpha: 1,
-      kCGWindowOwnerName: "Ontology Atlas",
-      kCGWindowBounds: { Width: 1280, Height: 820 },
-    },
-  ]);
-});
-
-test("validateWindowRequirements checks owner name and minimum size", () => {
-  const windows = [
-    {
-      kCGWindowOwnerName: "Ontology Atlas",
-      kCGWindowBounds: { Width: 1280, Height: 821 },
-    },
-  ];
-
-  assert.equal(
-    validateWindowRequirements(windows, {
-      requireOwnerName: "Ontology Atlas",
-      minWindowSize: { width: 1040, height: 720 },
-    }),
-    null,
-  );
-  assert.match(
-    validateWindowRequirements(windows, { requireOwnerName: "Other App" }),
-    /owner name/,
-  );
-  assert.match(
-    validateWindowRequirements(windows, { minWindowSize: { width: 1600, height: 900 } }),
-    /at least 1600x900/,
-  );
-});
-
-test("windowCaptureTargets keeps CoreGraphics window ids for screenshot capture", () => {
-  assert.deepEqual(
-    windowCaptureTargets([
-      {
-        kCGWindowNumber: 68525,
-        kCGWindowOwnerPID: 101,
-        kCGWindowOwnerName: "Ontology Atlas",
-        kCGWindowName: "Ontology Atlas",
-      },
-      {
-        kCGWindowOwnerPID: 101,
-        kCGWindowOwnerName: "Ontology Atlas",
-      },
-    ]),
-    [
-      {
-        id: 68525,
-        ownerPid: 101,
-        ownerName: "Ontology Atlas",
-        name: "Ontology Atlas",
-        bounds: null,
-        alpha: null,
-        sharingState: null,
-        storeType: null,
-        memoryUsage: null,
-      },
-    ],
-  );
-});
-
-test("validateCapturableWindowRows requires at least one successful window capture", () => {
-  assert.equal(
-    validateCapturableWindowRows([
-      { id: 1, ownerName: "Ontology Atlas", ok: false, stderr: "could not create image from window" },
-      { id: 2, ownerName: "Ontology Atlas", ok: true, method: "bounds-region", stderr: "", bytes: 2048 },
-    ]),
-    null,
-  );
-  assert.match(validateCapturableWindowRows([]), /no CoreGraphics window ids/);
-  assert.match(
-    validateCapturableWindowRows([
-      { id: 1, ownerName: "Ontology Atlas", ok: false, stderr: "could not create image from window" },
-    ]),
-    /could not create image from window/,
-  );
-});
-
-test("classifyVisualEvidenceBlocker explains foreground and blank-capture failures", () => {
-  assert.equal(
-    classifyVisualEvidenceBlocker({
-      activation: {
-        frontmost: false,
-        stderr: "post-activation Accessibility probe timed out after 3000ms",
-      },
-      captureRows: [
-        { ok: false, stderr: "window-id: could not create image from window", artifactPath: null },
-        {
-          ok: false,
-          stderr: "full-screen: image appears blank or black (nonDarkRatio 0)",
-          artifactPath: null,
-        },
-      ],
-    }),
-    "macos-automation-and-screen-capture-blocked",
-  );
-  assert.equal(
-    classifyVisualEvidenceBlocker({
-      activation: { frontmost: false },
-      captureRows: [
-        { ok: false, stderr: "window-id: could not create image from window", artifactPath: null },
-      ],
-    }),
-    "foreground-activation-unconfirmed",
-  );
-  assert.equal(
-    classifyVisualEvidenceBlocker({
-      activation: { frontmost: true },
-      captureRows: [
-        { ok: false, stderr: "full-screen: image appears blank or black (nonDarkRatio 0)", artifactPath: null },
-      ],
-    }),
-    "screen-capture-returned-blank-image",
-  );
-  assert.equal(
-    classifyVisualEvidenceBlocker({
-      activation: { frontmost: true },
-      captureRows: [
-        { ok: true, stderr: "", artifactPath: "/tmp/ontology-atlas.png" },
-      ],
-    }),
-    "captured",
-  );
-});
-
-test("visualEvidenceBlockerHint gives actionable remediation for macOS automation blockers", () => {
-  assert.deepEqual(
-    visualEvidenceBlockerHint("macos-automation-and-screen-capture-blocked"),
-    {
-      summary:
-        "macOS automation and screen capture blocked visual evidence; WebView proof may still be valid.",
-      nextActions: [
-        "Grant Accessibility permission to the terminal or Codex host running the verifier.",
-        "Grant Screen Recording permission, then rerun with --try-window-screenshot or --require-capturable-window.",
-        "Use the saved WebView evidence JSON as deterministic route proof until PNG capture is available.",
-      ],
-    },
-  );
-});
-
-test("formatVisualEvidenceHandoffLines prints blocker summary and next actions", () => {
-  assert.deepEqual(
-    formatVisualEvidenceHandoffLines({
-      blocker: "macos-automation-and-screen-capture-blocked",
-      requestedPath: "/tmp/ontology-atlas.png",
-      diagnosticsPath: "/tmp/ontology-atlas.png.diagnostics.json",
-      webviewEvidencePath: "/tmp/ontology-atlas.webview.json",
-      hint: visualEvidenceBlockerHint("macos-automation-and-screen-capture-blocked"),
-    }),
-    [
-      "[desktop-app-verify:visual-evidence] blocker macos-automation-and-screen-capture-blocked: macOS automation and screen capture blocked visual evidence; WebView proof may still be valid.",
-      "[desktop-app-verify:visual-evidence] WebView route proof: /tmp/ontology-atlas.webview.json",
-      "[desktop-app-verify:visual-evidence] next action 1: Grant Accessibility permission to the terminal or Codex host running the verifier.",
-      "[desktop-app-verify:visual-evidence] next action 2: Grant Screen Recording permission, then rerun with --try-window-screenshot or --require-capturable-window.",
-      "[desktop-app-verify:visual-evidence] next action 3: Use the saved WebView evidence JSON as deterministic route proof until PNG capture is available.",
-      "[desktop-app-verify:visual-evidence] diagnostics saved /tmp/ontology-atlas.png.diagnostics.json",
-      "[desktop-app-verify:visual-evidence] screenshot unavailable for /tmp/ontology-atlas.png",
-    ],
-  );
-});
-
-test("formatWindowDiagnosticsPayload includes capture and Accessibility evidence", () => {
-  assert.deepEqual(
-    formatWindowDiagnosticsPayload({
-      pids: [101],
-      windows: [
-        {
-          kCGWindowNumber: 81157,
-          kCGWindowOwnerPID: 101,
-          kCGWindowOwnerName: "Ontology Atlas",
-          kCGWindowName: "Ontology Atlas",
-          kCGWindowBounds: { X: 116, Y: 98, Width: 1280, Height: 821 },
-          kCGWindowLayer: 0,
-          kCGWindowIsOnscreen: true,
-          kCGWindowAlpha: 1,
-          kCGWindowSharingState: 1,
-          kCGWindowStoreType: 2,
-          kCGWindowMemoryUsage: 4096,
-        },
-      ],
-      accessibilityRows: [
-        {
-          pid: 101,
-          processName: "ontology-atlas",
-          frontmost: false,
-          windowCount: 0,
-          uiElementCount: 2,
-        },
-      ],
-      captureRows: [
-        {
-          id: 81157,
-          ownerName: "Ontology Atlas",
-          sharingState: 1,
-          alpha: 1,
-          ok: false,
-          method: "bounds-region",
-          stderr: "window-id: could not create image from window; bounds-region: could not create image from rect",
-          bytes: 0,
-        },
-      ],
-    }),
-    {
-      pids: [101],
-      windows: [
-        {
-          windowNumber: 81157,
-          ownerPid: 101,
-          ownerName: "Ontology Atlas",
-          name: "Ontology Atlas",
-          bounds: { X: 116, Y: 98, Width: 1280, Height: 821 },
-          layer: 0,
-          onscreen: true,
-          alpha: 1,
-          sharingState: 1,
-          storeType: 2,
-          memoryUsage: 4096,
-        },
-      ],
-      accessibilityRows: [
-        {
-          pid: 101,
-          processName: "ontology-atlas",
-          frontmost: false,
-          windowCount: 0,
-          uiElementCount: 2,
-        },
-      ],
-      captureRows: [
-        {
-          windowNumber: 81157,
-          ownerName: "Ontology Atlas",
-          sharingState: 1,
-          alpha: 1,
-          ok: false,
-          method: "bounds-region",
-          stderr: "window-id: could not create image from window; bounds-region: could not create image from rect",
-          bytes: 0,
-          artifactPath: null,
-        },
-      ],
-    },
-  );
-});
-
-test("collectWindowDiagnostics can keep best-effort visual diagnostics when Accessibility is unavailable", () => {
-  assert.deepEqual(
-    collectWindowDiagnostics({
-      executablePath: "/tmp/Ontology Atlas.app/Contents/MacOS/ontology-atlas",
-      processIdsFn: () => [101],
-      readOnscreenWindowsFn: () => "[]",
-      readAccessibilityWindowsFn: () => {
-        throw new Error("System Events did not respond within 3000ms");
-      },
-      allowAccessibilityFailure: true,
-      captureRows: [
-        {
-          id: 81157,
-          ownerName: "Ontology Atlas",
-          sharingState: 1,
-          alpha: 1,
-          ok: false,
-          method: "bounds-region",
-          stderr: "window-id: could not create image from window",
-          bytes: 0,
-        },
-      ],
-    }),
-    {
-      pids: [101],
-      windows: [],
-      accessibilityRows: [],
-      accessibilityError: "System Events did not respond within 3000ms",
-      captureRows: [
-        {
-          windowNumber: 81157,
-          ownerName: "Ontology Atlas",
-          sharingState: 1,
-          alpha: 1,
-          ok: false,
-          method: "bounds-region",
-          stderr: "window-id: could not create image from window",
-          bytes: 0,
-          artifactPath: null,
-        },
-      ],
-    },
-  );
-});
-
-test("formatWindowDiagnosticsPayload records full-screen visual evidence fallback rows", () => {
-  assert.deepEqual(
-    formatWindowDiagnosticsPayload({
-      pids: [101],
-      windows: [],
-      accessibilityRows: [],
-      captureRows: [
-        {
-          id: null,
-          ownerName: "desktop",
-          sharingState: null,
-          alpha: null,
-          ok: true,
-          method: "full-screen",
-          stderr: "",
-          bytes: 4096,
-          artifactPath: "/tmp/ontology-atlas-full-screen.png",
-        },
-      ],
-    }).captureRows,
-    [
-      {
-        windowNumber: null,
-        ownerName: "desktop",
-        sharingState: null,
-        alpha: null,
-        ok: true,
-        method: "full-screen",
-        stderr: "",
-        bytes: 4096,
-        artifactPath: "/tmp/ontology-atlas-full-screen.png",
-      },
-    ],
-  );
-});
-
-test("validateVisualEvidenceStats rejects blank or low-contrast screenshots", () => {
-  assert.equal(
-    validateVisualEvidenceStats({
-      width: 3024,
-      height: 1964,
-      sampleCount: 4096,
-      nonDarkRatio: 0.02,
-      lumaSpread: 32,
-    }),
-    null,
-  );
-  assert.match(
-    validateVisualEvidenceStats({
-      width: 3024,
-      height: 1964,
-      sampleCount: 4096,
-      nonDarkRatio: 0,
-      lumaSpread: 0,
-    }),
-    /blank or black/,
-  );
-  assert.match(
-    validateVisualEvidenceStats({
-      width: 3024,
-      height: 1964,
-      sampleCount: 4096,
-      nonDarkRatio: 0.02,
-      lumaSpread: 2,
-    }),
-    /too little visible contrast/,
-  );
-});
-
-test("Accessibility window probe targets launched process ids", () => {
-  const script = buildAccessibilityWindowProbeScript([101, 202]);
-
-  assert.match(script, /procPid = 101 or procPid = 202/);
-  assert.match(script, /count of windows of proc/);
-});
-
-test("foreground activation targets both bundle id and launched process ids", () => {
-  const script = buildForegroundActivationScript({
-    bundleIdentifier: "dev.jinan.ontology-atlas",
-    pids: [101, 202],
-  });
-
-  assert.match(script, /tell application id "dev\.jinan\.ontology-atlas" to activate/);
-  assert.match(script, /procPid = 101 or procPid = 202/);
-  assert.match(script, /set frontmost of proc to true/);
-  assert.match(script, /bundle=/);
-  assert.match(script, /pid=/);
-});
-
-test("parseAccessibilityWindowRows reads System Events tabular output", () => {
-  assert.deepEqual(
-    parseAccessibilityWindowRows(
-      "101\tOntology Atlas\tfalse\t1\t3\n202\tOther\ttrue\t0\t2\n",
-    ),
-    [
-      {
-        pid: 101,
-        processName: "Ontology Atlas",
-        frontmost: false,
-        windowCount: 1,
-        uiElementCount: 3,
-      },
-      {
-        pid: 202,
-        processName: "Other",
-        frontmost: true,
-        windowCount: 0,
-        uiElementCount: 2,
-      },
-    ],
-  );
-});
-
-test("validateAccessibilityWindowRows requires System Events windows", () => {
-  assert.equal(
-    validateAccessibilityWindowRows([
-      {
-        pid: 101,
-        processName: "Ontology Atlas",
-        frontmost: false,
-        windowCount: 1,
-        uiElementCount: 3,
-      },
-    ]),
-    null,
-  );
-  assert.match(
-    validateAccessibilityWindowRows([
-      {
-        pid: 101,
-        processName: "ontology-atlas",
-        frontmost: false,
-        windowCount: 0,
-        uiElementCount: 2,
-      },
-    ]),
-    /no Accessibility windows/,
-  );
-  assert.match(validateAccessibilityWindowRows([]), /did not find/);
-  assert.match(
-    validateAccessibilityWindowRows([
-      {
-        pid: 101,
-        processName: "Ontology Atlas",
-        frontmost: false,
-        windowCount: 0,
-        uiElementCount: 0,
-      },
-    ]),
-    /no Accessibility windows/,
-  );
-});
-
-test("validateFrontmostAccessibilityRows requires a foreground launched process", () => {
-  assert.equal(
-    validateFrontmostAccessibilityRows([
-      {
-        pid: 101,
-        processName: "ontology-atlas",
-        frontmost: true,
-        windowCount: 0,
-        uiElementCount: 2,
-      },
-    ]),
-    null,
-  );
-  assert.match(validateFrontmostAccessibilityRows([]), /did not find/);
-  assert.match(
-    validateFrontmostAccessibilityRows([
-      {
-        pid: 101,
-        processName: "ontology-atlas",
-        frontmost: false,
-        windowCount: 0,
-        uiElementCount: 2,
-      },
-    ]),
-    /not frontmost/,
-  );
-});
-
-test("existingProcessPatterns match stale macOS app copies with the same executable", () => {
-  assert.deepEqual(
-    existingProcessPatterns({
-      appPath: "/Users/me/Ontology Atlas.app",
-      executablePath: "/Users/me/Ontology Atlas.app/Contents/MacOS/ontology-atlas",
-    }),
-    [
-      "/Users/me/Ontology Atlas\\.app/Contents/MacOS/ontology-atlas",
-      "\\.app/Contents/MacOS/ontology-atlas$",
-    ],
   );
 });
